Extract modeling utilities into package
