+++ demistifai/constants.py
@@ -1,0 +1,904 @@
+"""Application-wide constants and reusable configuration values."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import List
+
+# CSS assets used throughout the application.
+APP_THEME_CSS = """
+<style>
+:root {
+    --surface-radius: 24px;
+    --surface-border: rgba(15, 23, 42, 0.12);
+    --surface-shadow: 0 24px 48px rgba(15, 23, 42, 0.14);
+    --surface-gradient: linear-gradient(160deg, rgba(226, 232, 240, 0.9), rgba(255, 255, 255, 0.96));
+    --accent-primary: #1d4ed8;
+    --accent-muted: rgba(30, 64, 175, 0.16);
+}
+
+.stApp {
+    background: radial-gradient(circle at top left, rgba(148, 163, 184, 0.16), transparent 45%),
+        radial-gradient(circle at 85% 10%, rgba(96, 165, 250, 0.16), transparent 50%),
+        #f8fafc;
+    color: #0f172a;
+    font-family: "Inter", "Segoe UI", sans-serif;
+}
+
+[data-testid="stMainBlock"] {
+    padding-top: 2.5rem;
+    padding-bottom: 3rem;
+}
+
+.main .block-container {
+    max-width: 1200px;
+    padding-left: 1.5rem;
+    padding-right: 1.5rem;
+}
+
+.section-surface {
+    display: none;
+}
+
+.section-surface-block {
+    position: relative;
+    margin-bottom: 1.45rem;
+    border-radius: var(--surface-radius);
+    border: 1px solid var(--surface-border);
+    background: var(--surface-gradient);
+    box-shadow: var(--surface-shadow);
+    padding: 1.85rem 2.1rem;
+    overflow: hidden;
+    color: #0f172a;
+}
+
+.section-surface-block::before {
+    content: "";
+    position: absolute;
+    inset: 0;
+    background: linear-gradient(150deg, rgba(59, 130, 246, 0.15), transparent 65%);
+    opacity: 0;
+    transition: opacity 0.3s ease;
+    pointer-events: none;
+}
+
+.section-surface-block:hover::before {
+    opacity: 1;
+}
+
+.section-surface-block > [data-testid="stElementContainer"] {
+    margin-bottom: 0.9rem;
+}
+
+.section-surface-block > [data-testid="stElementContainer"]:first-child,
+.section-surface-block > [data-testid="stElementContainer"]:last-child {
+    margin-bottom: 0;
+}
+
+.section-surface-block h2,
+.section-surface-block h3,
+.section-surface-block h4 {
+    margin-top: 0;
+    color: inherit;
+}
+
+.section-surface-block p,
+.section-surface-block ul {
+    font-size: 0.98rem;
+    line-height: 1.65;
+    color: inherit;
+}
+
+.section-surface-block ul {
+    padding-left: 1.25rem;
+}
+
+.section-surface-block .section-caption {
+    display: inline-flex;
+    align-items: center;
+    gap: 0.45rem;
+    font-size: 0.9rem;
+    text-transform: uppercase;
+    letter-spacing: 0.09em;
+    color: rgba(15, 23, 42, 0.65);
+}
+
+.section-surface-block .section-caption span {
+    display: inline-flex;
+    width: 40px;
+    height: 2px;
+    background: rgba(15, 23, 42, 0.12);
+}
+
+.section-surface-block.section-surface--hero {
+    padding: 2.6rem 2.8rem;
+    background: linear-gradient(160deg, #1d4ed8, #312e81);
+    color: #f8fafc;
+    border: 1px solid rgba(255, 255, 255, 0.28);
+    box-shadow: 0 28px 70px rgba(30, 64, 175, 0.35);
+}
+
+.section-surface-block.section-surface--hero::before {
+    display: none;
+}
+
+.section-surface-block.section-surface--hero .section-caption {
+    color: rgba(241, 245, 249, 0.85);
+}
+
+.section-surface-block.section-surface--hero .section-caption span {
+    background: rgba(241, 245, 249, 0.35);
+}
+
+.section-surface-block .surface-columns {
+    display: grid;
+    gap: 1.4rem;
+}
+
+.section-surface-block .surface-columns.two {
+    grid-template-columns: repeat(auto-fit, minmax(240px, 1fr));
+}
+
+.section-surface-block .surface-columns.three {
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+}
+
+.section-surface--hero [data-testid="column"]:nth-child(2) > div:first-of-type {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.section-surface--hero [data-testid="column"]:nth-child(2) > div:first-of-type > div[data-testid="stVerticalBlock"] {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 1.6rem;
+    width: 100%;
+}
+
+.section-surface--hero [data-testid="column"]:nth-child(2) > div:first-of-type > div[data-testid="stVerticalBlock"] > div {
+    width: 100%;
+}
+
+.section-surface--hero [data-testid="column"]:nth-child(2) > div:first-of-type > div[data-testid="stVerticalBlock"] > div[data-testid="stButton"] {
+    display: flex;
+    justify-content: center;
+}
+
+.section-surface--hero [data-testid="column"]:nth-child(2) > div:first-of-type > div[data-testid="stVerticalBlock"] > div[data-testid="stButton"] > button {
+    width: 100%;
+    max-width: 260px;
+}
+
+.section-surface--hero [data-testid="column"]:nth-child(2) .hero-info-grid {
+    width: 100%;
+    justify-items: center;
+    gap: 1.6rem;
+}
+
+.section-surface--hero [data-testid="column"]:nth-child(2) .hero-info-card {
+    text-align: center;
+    margin: 0 auto;
+    max-width: 360px;
+}
+
+.section-surface--hero [data-testid="column"]:nth-child(2) .hero-info-card h3,
+.section-surface--hero [data-testid="column"]:nth-child(2) .hero-info-card p {
+    text-align: center;
+}
+
+.section-surface--hero [data-testid="column"]:nth-child(2) [data-testid="stButton"] {
+    width: 100%;
+}
+
+.section-surface--hero [data-testid="column"]:nth-child(2) [data-testid="stButton"] > button {
+    max-width: 260px;
+    margin: 0 auto;
+}
+
+.callout {
+    position: relative;
+    border-radius: 18px;
+    border: 1px solid var(--surface-border);
+    background: rgba(255, 255, 255, 0.92);
+    box-shadow: 0 20px 40px rgba(15, 23, 42, 0.08);
+    padding: 1.35rem 1.5rem;
+    backdrop-filter: blur(6px);
+}
+
+.callout h4,
+.callout h5 {
+    margin: 0 0 0.65rem 0;
+    font-weight: 600;
+    color: inherit;
+}
+
+.callout p {
+    margin: 0;
+    font-size: 0.98rem;
+    line-height: 1.6;
+}
+
+.callout--mission {
+    background: linear-gradient(145deg, rgba(59, 130, 246, 0.12), rgba(191, 219, 254, 0.35));
+    border-color: rgba(37, 99, 235, 0.3);
+}
+
+.callout--info {
+    background: linear-gradient(150deg, rgba(191, 219, 254, 0.35), rgba(219, 234, 254, 0.65));
+    border-color: rgba(37, 99, 235, 0.25);
+}
+
+.callout-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
+    gap: 1rem;
+    margin-top: 0.9rem;
+}
+
+.callout--outcome {
+    display: flex;
+    gap: 0.85rem;
+    align-items: flex-start;
+    border-color: rgba(15, 23, 42, 0.08);
+}
+
+.callout-icon {
+    flex-shrink: 0;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    width: 44px;
+    height: 44px;
+    border-radius: 14px;
+    background: rgba(37, 99, 235, 0.12);
+    font-size: 1.35rem;
+}
+
+.callout-body h5 {
+    margin-bottom: 0.35rem;
+}
+
+.callout-body p {
+    font-size: 0.95rem;
+    color: rgba(15, 23, 42, 0.82);
+}
+
+.mission-preview-stack {
+    display: grid;
+    gap: 1.1rem;
+}
+
+.mission-card,
+.inbox-preview-card {
+    position: relative;
+    border-radius: 20px;
+    border: 1px solid rgba(37, 99, 235, 0.18);
+    padding: 1.35rem 1.5rem;
+    box-shadow: 0 22px 44px rgba(15, 23, 42, 0.12);
+}
+
+.mission-card {
+    background: linear-gradient(150deg, rgba(191, 219, 254, 0.45), rgba(59, 130, 246, 0.2));
+}
+
+.inbox-preview-card {
+    background: linear-gradient(150deg, rgba(248, 250, 252, 0.9), rgba(191, 219, 254, 0.55));
+    border-color: rgba(37, 99, 235, 0.16);
+}
+
+.mission-card::before,
+.inbox-preview-card::before {
+    content: "";
+    position: absolute;
+    inset: 0;
+    border-radius: inherit;
+    border: 1px solid rgba(255, 255, 255, 0.35);
+    opacity: 0.45;
+    pointer-events: none;
+}
+
+.mission-header,
+.preview-header {
+    display: flex;
+    align-items: center;
+    gap: 0.85rem;
+}
+
+.mission-header-icon,
+.preview-header-icon {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    width: 48px;
+    height: 48px;
+    border-radius: 16px;
+    background: rgba(37, 99, 235, 0.18);
+    font-size: 1.5rem;
+}
+
+.mission-header h4,
+.preview-header h4 {
+    margin: 0;
+    font-size: 1.15rem;
+}
+
+.mission-header p,
+.preview-header p {
+    margin: 0.2rem 0 0 0;
+    font-size: 0.95rem;
+    color: rgba(15, 23, 42, 0.75);
+}
+
+.mission-points {
+    margin: 1rem 0 0 0;
+    padding-left: 1.1rem;
+    font-size: 0.96rem;
+    line-height: 1.6;
+    color: rgba(15, 23, 42, 0.86);
+}
+
+.preview-note {
+    margin: 0.9rem 0 0 0;
+    font-size: 0.86rem;
+    color: rgba(15, 23, 42, 0.65);
+}
+
+.nerd-toggle-card {
+    display: contents;
+    grid-template-columns: minmax(0, 1fr) auto;
+    align-items: center;
+    gap: 1.1rem;
+    padding: 1.1rem 1.1rem;
+    border-radius: 18px;
+    border: 1px solid rgba(59, 130, 246, 0.28);
+    background: linear-gradient(150deg, rgba(37, 99, 235, 0.12), rgba(191, 219, 254, 0.22));
+    box-shadow: 0 16px 32px rgba(15, 23, 42, 0.12);
+    margin-bottom: 0.8rem;
+}
+
+.nerd-toggle-title {
+    font-size: 1.05rem;
+    font-weight: 600;
+    color: #0f172a;
+    display: flex;
+    align-items: center;
+    gap: 0.45rem;
+}
+
+.nerd-toggle-description {
+    font-size: 0.9rem;
+    color: #334155;
+    margin-top: 0.2rem;
+}
+
+.nerd-toggle-card label[data-testid="stToggle"] {
+    display: flex;
+    justify-content: flex-end;
+}
+
+.nerd-toggle-card label[data-testid="stToggle"] > div[role="switch"] {
+    background: rgba(148, 163, 184, 0.45);
+    box-shadow: inset 0 0 0 1px rgba(15, 23, 42, 0.12);
+}
+
+.nerd-toggle-card label[data-testid="stToggle"] > div[role="switch"][aria-checked="true"] {
+    background: linear-gradient(120deg, rgba(37, 99, 235, 0.85), rgba(59, 130, 246, 0.75));
+    box-shadow: inset 0 0 0 1px rgba(255, 255, 255, 0.35);
+}
+
+.info-metric-grid {
+    display: grid;
+    gap: 0.9rem;
+}
+
+.info-metric-card {
+    padding: 0.95rem 1.1rem;
+    border-radius: 16px;
+    background: linear-gradient(145deg, rgba(15, 23, 42, 0.85), rgba(15, 23, 42, 0.75));
+    color: #e2e8f0;
+    box-shadow: 0 18px 36px rgba(15, 23, 42, 0.22);
+}
+
+.info-metric-card .label {
+    font-size: 0.8rem;
+    text-transform: uppercase;
+    letter-spacing: 0.08em;
+    opacity: 0.8;
+}
+
+.info-metric-card .value {
+    font-size: 1.5rem;
+    font-weight: 700;
+    margin-top: 0.35rem;
+}
+
+[data-testid="stSidebar"] {
+    background: linear-gradient(180deg, rgba(30, 41, 59, 0.92), rgba(30, 41, 59, 0.85));
+    border-right: 1px solid rgba(30, 41, 59, 0.35);
+    color: #f8fafc;
+}
+
+[data-testid="stSidebar"] > div:first-child {
+    padding: 2.6rem 1.6rem 2.1rem;
+}
+
+.sidebar-shell {
+    display: flex;
+    flex-direction: column;
+    gap: 2rem;
+}
+
+.sidebar-brand {
+    display: flex;
+    flex-direction: column;
+    gap: 0.4rem;
+}
+
+.sidebar-brand .sidebar-title {
+    margin: 0;
+    font-size: 1.18rem;
+    font-weight: 700;
+    color: #1d4ed8;
+}
+
+.sidebar-brand .sidebar-subtitle {
+    margin: 0;
+    color: rgba(248, 250, 252, 0.8);
+    line-height: 1.6;
+    font-size: 0.95rem;
+}
+
+.sidebar-panel {
+    display: grid;
+    gap: 1.2rem;
+}
+
+.sidebar-panel h4 {
+    margin: 0;
+    font-size: 1rem;
+    color: rgba(248, 250, 252, 0.85);
+}
+
+.sidebar-progress {
+    display: flex;
+    flex-direction: column;
+    gap: 0.6rem;
+}
+
+.sidebar-stage-card {
+    display: grid;
+    grid-template-columns: auto minmax(0, 1fr);
+    gap: 0.75rem;
+    align-items: center;
+    padding: 0.8rem 0.9rem;
+    border-radius: 16px;
+    background: rgba(30, 41, 59, 0.5);
+    border: 1px solid rgba(148, 163, 184, 0.18);
+    box-shadow: 0 20px 42px rgba(15, 23, 42, 0.16);
+}
+
+.sidebar-stage-card__icon {
+    display: grid;
+    place-items: center;
+    width: 48px;
+    height: 48px;
+    border-radius: 16px;
+    background: rgba(59, 130, 246, 0.14);
+    color: #1d4ed8;
+    font-size: 1.3rem;
+    box-shadow: inset 0 0 0 1px rgba(59, 130, 246, 0.28);
+}
+
+.sidebar-stage-card__meta {
+    display: flex;
+    flex-direction: column;
+    gap: 0.2rem;
+}
+
+.sidebar-stage-card__eyebrow {
+    font-size: 0.7rem;
+    font-weight: 700;
+    letter-spacing: 0.1em;
+    text-transform: uppercase;
+    color: rgba(248, 250, 252, 0.65);
+}
+
+.sidebar-stage-card__title {
+    margin: 0;
+    font-size: 1.02rem;
+    font-weight: 700;
+    color: #f8fafc;
+}
+
+.sidebar-stage-card__description {
+    margin: 0;
+    color: rgba(248, 250, 252, 0.75);
+    line-height: 1.55;
+    font-size: 0.9rem;
+}
+
+[data-testid="stSidebar"] .stToggle {
+    padding: 0.5rem 0.25rem 0.2rem;
+}
+
+[data-testid="stSidebar"] .stButton > button {
+    width: 100%;
+    border-radius: 14px;
+    border: none;
+    background: linear-gradient(140deg, #1d4ed8, #312e81);
+    color: #f8fafc;
+    font-weight: 600;
+    box-shadow: 0 18px 42px rgba(30, 64, 175, 0.28);
+}
+
+[data-testid="stSidebar"] .stButton > button:hover {
+    background: linear-gradient(140deg, #1e40af, #1d4ed8);
+    box-shadow: 0 22px 48px rgba(30, 64, 175, 0.32);
+}
+
+[data-testid="stSidebar"] .stButton > button:focus-visible {
+    outline: 3px solid rgba(59, 130, 246, 0.55);
+    outline-offset: 2px;
+}
+
+.metric-highlight {
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+    padding: 0.85rem 1.1rem;
+    border-radius: 16px;
+    background: rgba(52, 211, 153, 0.14);
+    border: 1px solid rgba(52, 211, 153, 0.28);
+    color: #047857;
+    font-weight: 600;
+    margin: 1rem 0;
+}
+
+.metric-highlight svg {
+    width: 28px;
+    height: 28px;
+}
+
+.pill-group {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.6rem;
+    margin: 1rem 0;
+}
+
+.pill-group .pill {
+    padding: 0.45rem 0.9rem;
+    border-radius: 999px;
+    background: rgba(59, 130, 246, 0.12);
+    color: #1d4ed8;
+    font-size: 0.85rem;
+    font-weight: 600;
+}
+</style>
+"""
+
+STAGE_TEMPLATE_CSS = """
+<style>
+:root {
+    --stage-card-radius: 18px;
+    --stage-card-border: rgba(15, 23, 42, 0.08);
+    --stage-card-shadow: 0 18px 40px rgba(15, 23, 42, 0.10);
+}
+
+.stage-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
+    gap: 1.2rem;
+    margin-top: 1.2rem;
+}
+
+.stage-progress-grid {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.6rem;
+    margin: 1rem 0 0;
+    padding: 0.5rem 0.75rem;
+    border-radius: 16px;
+    background: rgba(15, 23, 42, 0.03);
+    border: 1px solid rgba(148, 163, 184, 0.24);
+    overflow-x: auto;
+}
+
+.stage-progress-grid::-webkit-scrollbar {
+    height: 6px;
+}
+
+.stage-progress-grid::-webkit-scrollbar-thumb {
+    background: rgba(100, 116, 139, 0.35);
+    border-radius: 999px;
+}
+
+.stage-card {
+    background: linear-gradient(145deg, rgba(255, 255, 255, 0.96), rgba(241, 245, 255, 0.9));
+    border-radius: var(--stage-card-radius);
+    border: 1px solid var(--stage-card-border);
+    box-shadow: var(--stage-card-shadow);
+    padding: 1.15rem 1.2rem;
+    transition: transform 0.2s ease, box-shadow 0.2s ease, border-color 0.2s ease;
+    color: #0f172a;
+    position: relative;
+    overflow: hidden;
+    display: block;
+    text-decoration: none;
+}
+
+.stage-card:link,
+.stage-card:visited {
+    color: #0f172a;
+    text-decoration: none;
+}
+
+.stage-card::after {
+    content: "";
+    position: absolute;
+    inset: 0;
+    background: radial-gradient(circle at top right, rgba(74, 108, 247, 0.18), transparent 55%);
+    opacity: 0;
+    transition: opacity 0.2s ease;
+}
+
+.stage-card:hover {
+    transform: translateY(-3px);
+    box-shadow: 0 22px 50px rgba(15, 23, 42, 0.14);
+}
+
+.stage-card:hover::after {
+    opacity: 1;
+}
+
+.stage-card:focus-visible {
+    outline: 3px solid rgba(74, 108, 247, 0.45);
+    outline-offset: 3px;
+}
+
+.stage-card .stage-icon {
+    font-size: 1.85rem;
+    line-height: 1;
+    margin-bottom: 0.35rem;
+}
+
+.stage-card .stage-title {
+    font-size: 1.05rem;
+    font-weight: 600;
+}
+
+.stage-card .stage-summary {
+    margin-top: 0.55rem;
+    font-size: 0.92rem;
+    line-height: 1.5;
+    color: #334155;
+}
+
+.stage-card.active {
+    border-color: rgba(74, 108, 247, 0.55);
+    background: linear-gradient(155deg, rgba(74, 108, 247, 0.18), rgba(241, 245, 255, 0.96));
+}
+
+.stage-card.active::after {
+    opacity: 1;
+}
+
+.stage-card__body {
+    display: grid;
+    gap: 0.65rem;
+}
+
+.stage-card .stage-meta {
+    display: flex;
+    flex-direction: column;
+    gap: 0.25rem;
+    font-size: 0.82rem;
+    color: rgba(15, 23, 42, 0.65);
+}
+
+.stage-card .stage-meta span {
+    display: inline-flex;
+    align-items: center;
+    gap: 0.45rem;
+}
+
+.stage-progress-chip {
+    display: inline-flex;
+    align-items: center;
+    gap: 0.4rem;
+    padding: 0.4rem 0.75rem;
+    border-radius: 999px;
+    background: rgba(59, 130, 246, 0.18);
+    color: #1d4ed8;
+    font-weight: 600;
+    font-size: 0.85rem;
+}
+</style>
+"""
+
+EMAIL_INBOX_TABLE_CSS = """
+<style>
+.email-inbox-wrapper {
+    border: 1px solid rgba(15, 23, 42, 0.12);
+    border-radius: 14px;
+    overflow: hidden;
+    box-shadow: 0 18px 40px rgba(15, 23, 42, 0.12);
+    background: linear-gradient(145deg, rgba(255, 255, 255, 0.96), rgba(241, 245, 255, 0.92));
+    margin: 0.75rem 0 1.1rem;
+}
+
+.email-inbox-header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0.65rem 1.1rem;
+    background: linear-gradient(120deg, rgba(30, 64, 175, 0.9), rgba(59, 130, 246, 0.85));
+    color: #f8fafc;
+    font-weight: 600;
+    font-size: 0.94rem;
+}
+
+.email-inbox-table {
+    width: 100%;
+    border-collapse: collapse;
+    font-size: 0.9rem;
+    color: #0f172a;
+}
+
+.email-inbox-table thead {
+    background: rgba(226, 232, 240, 0.75);
+    text-transform: uppercase;
+    letter-spacing: 0.04em;
+    font-size: 0.76rem;
+    color: #1e293b;
+}
+
+.email-inbox-table th,
+.email-inbox-table td {
+    padding: 0.75rem 1rem;
+    border-bottom: 1px solid rgba(148, 163, 184, 0.35);
+    vertical-align: top;
+}
+
+.email-inbox-table tbody tr:hover {
+    background: rgba(191, 219, 254, 0.45);
+}
+
+.email-inbox-table tbody tr:nth-child(even) {
+    background: rgba(248, 250, 252, 0.85);
+}
+
+.email-inbox-empty {
+    padding: 1rem 1.25rem;
+    color: #475569;
+    font-style: italic;
+}
+</style>
+"""
+
+LIFECYCLE_CYCLE_CSS = """
+<style>
+.lifecycle-cycle {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-wrap: wrap;
+    gap: 0.9rem;
+    margin: 1.2rem 0 1.6rem;
+}
+
+.lifecycle-cycle .cycle-node {
+    width: 118px;
+    height: 118px;
+    border-radius: 50%;
+    background: linear-gradient(160deg, rgba(59, 130, 246, 0.15), rgba(59, 130, 246, 0.05));
+    border: 2px solid rgba(37, 99, 235, 0.35);
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    padding: 1rem;
+    gap: 0.4rem;
+    color: #1d4ed8;
+    font-weight: 600;
+    box-shadow: 0 22px 48px rgba(37, 99, 235, 0.18);
+}
+
+.lifecycle-cycle .cycle-node .cycle-icon {
+    font-size: 1.45rem;
+}
+
+.lifecycle-cycle .cycle-node .cycle-title {
+    font-size: 0.95rem;
+}
+
+.lifecycle-cycle .cycle-node .cycle-caption {
+    font-size: 0.75rem;
+    color: rgba(15, 23, 42, 0.65);
+}
+
+.lifecycle-cycle .cycle-arrow {
+    font-size: 1.65rem;
+    color: rgba(15, 23, 42, 0.45);
+}
+
+.lifecycle-cycle .cycle-loop {
+    width: 60px;
+    height: 60px;
+    border-radius: 50%;
+    background: rgba(59, 130, 246, 0.14);
+    display: grid;
+    place-items: center;
+    color: #1d4ed8;
+    font-size: 1.2rem;
+    box-shadow: 0 18px 36px rgba(59, 130, 246, 0.18);
+}
+</style>
+"""
+
+
+@dataclass(frozen=True)
+class StageMeta:
+    key: str
+    title: str
+    icon: str
+    summary: str
+    description: str
+
+
+STAGES: List[StageMeta] = [
+    StageMeta("intro", "Welcome", "🚀", "Kickoff", "Meet the mission and trigger the guided build."),
+    StageMeta(
+        "overview",
+        "Start your machine",
+        "🧭",
+        "See the journey",
+        "Tour the steps, set Nerd Mode, and align on what you'll do.",
+    ),
+    StageMeta("data", "Prepare Data", "📊", "Curate examples", "Inspect labeled emails and get the dataset ready for learning."),
+    StageMeta("train", "Train", "🧠", "Teach the model", "Configure the split and teach the model on your dataset."),
+    StageMeta("evaluate", "Evaluate", "🧪", "Check results", "Check metrics, inspect the confusion matrix, and stress-test."),
+    StageMeta("classify", "Use", "📬", "Route emails", "Route new messages, correct predictions, and adapt."),
+    StageMeta("model_card", "Model Card", "📄", "Document system", "Summarize performance, intended use, and governance."),
+]
+
+STAGE_INDEX = {stage.key: idx for idx, stage in enumerate(STAGES)}
+STAGE_BY_KEY = {stage.key: stage for stage in STAGES}
+
+CLASSES = ["spam", "safe"]
+AUTONOMY_LEVELS = [
+    "Moderate autonomy (recommendation)",
+    "High autonomy (auto-route)",
+]
+
+DATASET_SUSPICIOUS_TLDS = [
+    ".ru",
+    ".xyz",
+    ".top",
+    ".biz",
+    ".win",
+    ".loan",
+    ".live",
+    ".icu",
+    ".cn",
+    ".gq",
+]
+
+DATASET_LEGIT_DOMAINS = [
+    "intranet.corp",
+    "sharepoint.corp",
+    "confluence.corp",
+    "workday.corp",
+    "hr.corp",
+    "vpn.corp",
+    "it.corp",
+]
+
+BRANDS = ["DocuSign", "Office 365", "OneDrive", "Zoom", "Adobe", "Okta", "Teams"]
+
+COURIERS = ["DHL", "UPS", "FedEx", "PostNL", "PostNord"]
+
+URGENCY = ["URGENT", "ACTION REQUIRED", "FINAL NOTICE", "IMMEDIATE", "24H", "TODAY"]
