+++ streamlit_app.py
@@ -8,10 +8,9 @@ import string
 import hashlib
 import re
 from collections import Counter
-from dataclasses import dataclass
 from datetime import datetime
 from contextlib import contextmanager
-from typing import Any, Dict, List, Optional, Tuple, TypedDict
+from typing import Any, Dict, List, Optional, Tuple
 from urllib.parse import urlencode
 from uuid import uuid4
 
@@ -21,2439 +20,88 @@ import pandas as pd
 import streamlit as st
 import streamlit.components.v1 as components
 
-from sentence_transformers import SentenceTransformer
-from sklearn.linear_model import LogisticRegression
-from sklearn.metrics import accuracy_score, confusion_matrix, precision_recall_fscore_support
+from demistifai.constants import (
+    APP_THEME_CSS,
+    AUTONOMY_LEVELS,
+    BRANDS,
+    CLASSES,
+    COURIERS,
+    DATASET_LEGIT_DOMAINS,
+    DATASET_SUSPICIOUS_TLDS,
+    EMAIL_INBOX_TABLE_CSS,
+    LIFECYCLE_CYCLE_CSS,
+    STAGE_BY_KEY,
+    STAGE_INDEX,
+    STAGE_TEMPLATE_CSS,
+    STAGES,
+    URGENCY,
+    StageMeta,
+)
+from demistifai.dataset import (
+    ATTACHMENT_MIX_PRESETS,
+    ATTACHMENT_TYPES,
+    DEFAULT_ATTACHMENT_MIX,
+    DEFAULT_DATASET_CONFIG,
+    EDGE_CASE_TEMPLATES,
+    DatasetConfig,
+    STARTER_INCOMING,
+    SUSPICIOUS_TLD_SUFFIXES,
+    build_dataset_from_config,
+    compute_dataset_hash,
+    compute_dataset_summary,
+    dataset_delta_story,
+    dataset_summary_delta,
+    explain_config_change,
+    generate_labeled_dataset,
+    lint_dataset,
+    _caps_ratio,
+    _count_money_mentions,
+    _count_suspicious_links,
+    _has_suspicious_tld,
+)
+from demistifai.modeling import (
+    FEATURE_DISPLAY_NAMES,
+    FEATURE_ORDER,
+    FEATURE_PLAIN_LANGUAGE,
+    HybridEmbedFeatsLogReg,
+    _combine_text,
+    _fmt_delta,
+    _fmt_pct,
+    _pr_acc_cm,
+    _predict_proba_batch,
+    _y01,
+    assess_performance,
+    cache_train_embeddings,
+    combine_text,
+    compute_confusion,
+    compute_numeric_features,
+    df_confusion,
+    encode_texts,
+    extract_urls,
+    features_matrix,
+    get_domain_tld,
+    get_encoder,
+    get_nearest_training_examples,
+    make_after_eval_story,
+    make_after_training_story,
+    model_kind_string,
+    numeric_feature_contributions,
+    plot_threshold_curves,
+    predict_spam_probability,
+    threshold_presets,
+    top_token_importances,
+    verdict_label,
+)
+
+from sklearn.metrics import accuracy_score, confusion_matrix
 from sklearn.model_selection import train_test_split
-from sklearn.preprocessing import StandardScaler
 
 st.set_page_config(page_title="demistifAI", page_icon="📧", layout="wide")
 
-APP_THEME_CSS = """
-<style>
-:root {
-    --surface-radius: 24px;
-    --surface-border: rgba(15, 23, 42, 0.12);
-    --surface-shadow: 0 24px 48px rgba(15, 23, 42, 0.14);
-    --surface-gradient: linear-gradient(160deg, rgba(226, 232, 240, 0.9), rgba(255, 255, 255, 0.96));
-    --accent-primary: #1d4ed8;
-    --accent-muted: rgba(30, 64, 175, 0.16);
-}
-
-.stApp {
-    background: radial-gradient(circle at top left, rgba(148, 163, 184, 0.16), transparent 45%),
-        radial-gradient(circle at 85% 10%, rgba(96, 165, 250, 0.16), transparent 50%),
-        #f8fafc;
-    color: #0f172a;
-    font-family: "Inter", "Segoe UI", sans-serif;
-}
-
-[data-testid="stMainBlock"] {
-    padding-top: 2.5rem;
-    padding-bottom: 3rem;
-}
-
-.main .block-container {
-    max-width: 1200px;
-    padding-left: 1.5rem;
-    padding-right: 1.5rem;
-}
-
-.section-surface {
-    display: none;
-}
-
-.section-surface-block {
-    position: relative;
-    margin-bottom: 1.45rem;
-    border-radius: var(--surface-radius);
-    border: 1px solid var(--surface-border);
-    background: var(--surface-gradient);
-    box-shadow: var(--surface-shadow);
-    padding: 1.85rem 2.1rem;
-    overflow: hidden;
-    color: #0f172a;
-}
-
-.section-surface-block::before {
-    content: "";
-    position: absolute;
-    inset: 0;
-    background: linear-gradient(150deg, rgba(59, 130, 246, 0.15), transparent 65%);
-    opacity: 0;
-    transition: opacity 0.3s ease;
-    pointer-events: none;
-}
-
-.section-surface-block:hover::before {
-    opacity: 1;
-}
-
-.section-surface-block > [data-testid="stElementContainer"] {
-    margin-bottom: 0.9rem;
-}
-
-.section-surface-block > [data-testid="stElementContainer"]:first-child,
-.section-surface-block > [data-testid="stElementContainer"]:last-child {
-    margin-bottom: 0;
-}
-
-.section-surface-block h2,
-.section-surface-block h3,
-.section-surface-block h4 {
-    margin-top: 0;
-    color: inherit;
-}
-
-.section-surface-block p,
-.section-surface-block ul {
-    font-size: 0.98rem;
-    line-height: 1.65;
-    color: inherit;
-}
-
-.section-surface-block ul {
-    padding-left: 1.25rem;
-}
-
-.section-surface-block .section-caption {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.45rem;
-    font-size: 0.9rem;
-    text-transform: uppercase;
-    letter-spacing: 0.09em;
-    color: rgba(15, 23, 42, 0.65);
-}
-
-.section-surface-block .section-caption span {
-    display: inline-flex;
-    width: 40px;
-    height: 2px;
-    background: rgba(15, 23, 42, 0.12);
-}
-
-.section-surface-block.section-surface--hero {
-    padding: 2.6rem 2.8rem;
-    background: linear-gradient(160deg, #1d4ed8, #312e81);
-    color: #f8fafc;
-    border: 1px solid rgba(255, 255, 255, 0.28);
-    box-shadow: 0 28px 70px rgba(30, 64, 175, 0.35);
-}
-
-.section-surface-block.section-surface--hero::before {
-    display: none;
-}
-
-.section-surface-block.section-surface--hero .section-caption {
-    color: rgba(241, 245, 249, 0.85);
-}
-
-.section-surface-block.section-surface--hero .section-caption span {
-    background: rgba(241, 245, 249, 0.35);
-}
-
-.section-surface-block .surface-columns {
-    display: grid;
-    gap: 1.4rem;
-}
-
-.section-surface-block .surface-columns.two {
-    grid-template-columns: repeat(auto-fit, minmax(240px, 1fr));
-}
-
-.section-surface-block .surface-columns.three {
-    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-}
-
-
-.section-surface--hero [data-testid="column"]:nth-child(2) > div:first-of-type {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.section-surface--hero [data-testid="column"]:nth-child(2) > div:first-of-type > div[data-testid="stVerticalBlock"] {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 1.6rem;
-    width: 100%;
-}
-
-.section-surface--hero [data-testid="column"]:nth-child(2) > div:first-of-type > div[data-testid="stVerticalBlock"] > div {
-    width: 100%;
-}
-
-.section-surface--hero [data-testid="column"]:nth-child(2) > div:first-of-type > div[data-testid="stVerticalBlock"] > div[data-testid="stButton"] {
-    display: flex;
-    justify-content: center;
-}
-
-.section-surface--hero [data-testid="column"]:nth-child(2) > div:first-of-type > div[data-testid="stVerticalBlock"] > div[data-testid="stButton"] > button {
-    width: 100%;
-    max-width: 260px;
-}
-
-.section-surface--hero [data-testid="column"]:nth-child(2) .hero-info-grid {
-    width: 100%;
-    justify-items: center;
-    gap: 1.6rem;
-}
-
-.section-surface--hero [data-testid="column"]:nth-child(2) .hero-info-card {
-    text-align: center;
-    margin: 0 auto;
-    max-width: 360px;
-}
-
-.section-surface--hero [data-testid="column"]:nth-child(2) .hero-info-card h3,
-.section-surface--hero [data-testid="column"]:nth-child(2) .hero-info-card p {
-    text-align: center;
-}
-
-.section-surface--hero [data-testid="column"]:nth-child(2) [data-testid="stButton"] {
-    width: 100%;
-}
-
-.section-surface--hero [data-testid="column"]:nth-child(2) [data-testid="stButton"] > button {
-    max-width: 260px;
-    margin: 0 auto;
-}
-
-.callout {
-    position: relative;
-    border-radius: 18px;
-    border: 1px solid var(--surface-border);
-    background: rgba(255, 255, 255, 0.92);
-    box-shadow: 0 20px 40px rgba(15, 23, 42, 0.08);
-    padding: 1.35rem 1.5rem;
-    backdrop-filter: blur(6px);
-}
-
-.callout h4,
-.callout h5 {
-    margin: 0 0 0.65rem 0;
-    font-weight: 600;
-    color: inherit;
-}
-
-.callout p {
-    margin: 0;
-    font-size: 0.98rem;
-    line-height: 1.6;
-}
-
-.callout--mission {
-    background: linear-gradient(145deg, rgba(59, 130, 246, 0.12), rgba(191, 219, 254, 0.35));
-    border-color: rgba(37, 99, 235, 0.3);
-}
-
-.callout--info {
-    background: linear-gradient(150deg, rgba(191, 219, 254, 0.35), rgba(219, 234, 254, 0.65));
-    border-color: rgba(37, 99, 235, 0.25);
-}
-
-.callout-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
-    gap: 1rem;
-    margin-top: 0.9rem;
-}
-
-.callout--outcome {
-    display: flex;
-    gap: 0.85rem;
-    align-items: flex-start;
-    border-color: rgba(15, 23, 42, 0.08);
-}
-
-.callout-icon {
-    flex-shrink: 0;
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    width: 44px;
-    height: 44px;
-    border-radius: 14px;
-    background: rgba(37, 99, 235, 0.12);
-    font-size: 1.35rem;
-}
-
-.callout-body h5 {
-    margin-bottom: 0.35rem;
-}
-
-.callout-body p {
-    font-size: 0.95rem;
-    color: rgba(15, 23, 42, 0.82);
-}
-
-.mission-preview-stack {
-    display: grid;
-    gap: 1.1rem;
-}
-
-.mission-card,
-.inbox-preview-card {
-    position: relative;
-    border-radius: 20px;
-    border: 1px solid rgba(37, 99, 235, 0.18);
-    padding: 1.35rem 1.5rem;
-    box-shadow: 0 22px 44px rgba(15, 23, 42, 0.12);
-}
-
-.mission-card {
-    background: linear-gradient(150deg, rgba(191, 219, 254, 0.45), rgba(59, 130, 246, 0.2));
-}
-
-.inbox-preview-card {
-    background: linear-gradient(150deg, rgba(248, 250, 252, 0.9), rgba(191, 219, 254, 0.55));
-    border-color: rgba(37, 99, 235, 0.16);
-}
-
-.mission-card::before,
-.inbox-preview-card::before {
-    content: "";
-    position: absolute;
-    inset: 0;
-    border-radius: inherit;
-    border: 1px solid rgba(255, 255, 255, 0.35);
-    opacity: 0.45;
-    pointer-events: none;
-}
-
-.mission-header,
-.preview-header {
-    display: flex;
-    align-items: center;
-    gap: 0.85rem;
-}
-
-.mission-header-icon,
-.preview-header-icon {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    width: 48px;
-    height: 48px;
-    border-radius: 16px;
-    background: rgba(37, 99, 235, 0.18);
-    font-size: 1.5rem;
-}
-
-.mission-header h4,
-.preview-header h4 {
-    margin: 0;
-    font-size: 1.15rem;
-}
-
-.mission-header p,
-.preview-header p {
-    margin: 0.2rem 0 0 0;
-    font-size: 0.95rem;
-    color: rgba(15, 23, 42, 0.75);
-}
-
-.mission-points {
-    margin: 1rem 0 0 0;
-    padding-left: 1.1rem;
-    font-size: 0.96rem;
-    line-height: 1.6;
-    color: rgba(15, 23, 42, 0.86);
-}
-
-.preview-note {
-    margin: 0.9rem 0 0 0;
-    font-size: 0.86rem;
-    color: rgba(15, 23, 42, 0.65);
-}
-
-.nerd-toggle-card {
-    display: contents;
-    grid-template-columns: minmax(0, 1fr) auto;
-    align-items: center;
-    gap: 1.1rem;
-    padding: 1.1rem 1.1rem;
-    border-radius: 18px;
-    border: 1px solid rgba(59, 130, 246, 0.28);
-    background: linear-gradient(150deg, rgba(37, 99, 235, 0.12), rgba(191, 219, 254, 0.22));
-    box-shadow: 0 16px 32px rgba(15, 23, 42, 0.12);
-    margin-bottom: 0.8rem;
-}
-
-.nerd-toggle-title {
-    font-size: 1.05rem;
-    font-weight: 600;
-    color: #0f172a;
-    display: flex;
-    align-items: center;
-    gap: 0.45rem;
-}
-
-.nerd-toggle-description {
-    font-size: 0.9rem;
-    color: #334155;
-    margin-top: 0.2rem;
-}
-
-.nerd-toggle-card label[data-testid="stToggle"] {
-    display: flex;
-    justify-content: flex-end;
-}
-
-.nerd-toggle-card label[data-testid="stToggle"] > div[role="switch"] {
-    background: rgba(148, 163, 184, 0.45);
-    box-shadow: inset 0 0 0 1px rgba(15, 23, 42, 0.12);
-}
-
-.nerd-toggle-card label[data-testid="stToggle"] > div[role="switch"][aria-checked="true"] {
-    background: linear-gradient(120deg, rgba(37, 99, 235, 0.85), rgba(59, 130, 246, 0.75));
-    box-shadow: inset 0 0 0 1px rgba(255, 255, 255, 0.35);
-}
-
-.info-metric-grid {
-    display: grid;
-    gap: 0.9rem;
-}
-
-.info-metric-card {
-    padding: 0.95rem 1.1rem;
-    border-radius: 16px;
-    background: linear-gradient(145deg, rgba(15, 23, 42, 0.85), rgba(15, 23, 42, 0.75));
-    color: #e2e8f0;
-    box-shadow: 0 18px 36px rgba(15, 23, 42, 0.22);
-}
-
-.info-metric-card .label {
-    font-size: 0.8rem;
-    text-transform: uppercase;
-    letter-spacing: 0.08em;
-    opacity: 0.8;
-}
-
-.info-metric-card .value {
-    font-size: 1.5rem;
-    font-weight: 700;
-    margin-top: 0.35rem;
-}
-
-[data-testid="stSidebar"] {
-    background: linear-gradient(180deg, rgba(30, 41, 59, 0.92), rgba(30, 41, 59, 0.85));
-    color: #e2e8f0;
-}
-
-[data-testid="stSidebar"] h1,
-[data-testid="stSidebar"] h2,
-[data-testid="stSidebar"] h3,
-[data-testid="stSidebar"] p {
-    color: #0f172a;
-}
-
-[data-testid="stSidebar"]::-webkit-scrollbar-thumb {
-    background: rgba(148, 163, 184, 0.5);
-    border-radius: 999px;
-}
-
-[data-testid="stAlert"] {
-    border-radius: 18px !important;
-    padding: 1rem 1.25rem !important;
-    border: 1px solid rgba(148, 163, 184, 0.35) !important;
-    box-shadow: 0 18px 36px rgba(15, 23, 42, 0.14) !important;
-}
-
-.stButton>button,
-.stDownloadButton>button {
-    border-radius: 999px;
-    font-weight: 600;
-    letter-spacing: 0.01em;
-}
-
-.stTabs [data-baseweb="tab-list"] {
-    gap: 0.5rem;
-    padding: 0.35rem;
-    background: rgba(148, 163, 184, 0.22);
-    border-radius: 999px;
-    margin-bottom: 1.25rem;
-}
-
-.stTabs [data-baseweb="tab"] {
-    background: transparent;
-    border-radius: 999px;
-    padding: 0.55rem 1.2rem;
-    color: #1e293b;
-    font-weight: 600;
-    border: 1px solid transparent;
-    transition: background 0.2s ease, color 0.2s ease, box-shadow 0.2s ease, transform 0.2s ease;
-}
-
-.stTabs [data-baseweb="tab"]:hover {
-    background: rgba(255, 255, 255, 0.82);
-    box-shadow: 0 12px 28px rgba(15, 23, 42, 0.16);
-}
-
-.stTabs [data-baseweb="tab"][aria-selected="true"] {
-    background: linear-gradient(140deg, rgba(59, 130, 246, 0.22), rgba(37, 99, 235, 0.4));
-    color: #0f172a;
-    box-shadow: 0 18px 40px rgba(37, 99, 235, 0.28);
-    border-color: rgba(37, 99, 235, 0.35);
-    transform: translateY(-1px);
-}
-
-.stTabs [data-baseweb="tab"]:focus {
-    outline: none;
-    box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.35);
-}
-
-.stTabs [data-baseweb="tab-panel"] {
-    padding-top: 0.35rem;
-}
-
-.mailbox-summary {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
-
-.mailbox-summary-card {
-    border-radius: 18px;
-    padding: 1rem 1.2rem;
-    background: linear-gradient(160deg, rgba(255, 255, 255, 0.96), rgba(226, 232, 240, 0.75));
-    border: 1px solid rgba(148, 163, 184, 0.35);
-    box-shadow: 0 18px 38px rgba(15, 23, 42, 0.14);
-}
-
-.mailbox-summary-card .summary-title {
-    font-size: 0.82rem;
-    text-transform: uppercase;
-    letter-spacing: 0.08em;
-    color: rgba(15, 23, 42, 0.58);
-    margin-bottom: 0.35rem;
-}
-
-.mailbox-summary-card .summary-value {
-    font-size: 2rem;
-    font-weight: 700;
-    color: #0f172a;
-    line-height: 1.05;
-}
-
-.mailbox-summary-card .summary-hint {
-    margin-top: 0.6rem;
-    font-size: 0.9rem;
-    color: #475569;
-}
-
-.mailbox-summary-card .summary-rows {
-    display: flex;
-    flex-direction: column;
-    gap: 0.4rem;
-    margin-top: 0.35rem;
-}
-
-.mailbox-summary-card .summary-row {
-    display: flex;
-    justify-content: space-between;
-    font-size: 0.92rem;
-    color: #1e293b;
-}
-
-.mailbox-summary-card .summary-row span:last-child {
-    font-weight: 600;
-}
-
-.hero-info-grid {
-    display: grid;
-    grid-template-columns: minmax(0, 1fr);
-    gap: 1.1rem;
-}
-
-.hero-info-card {
-    position: relative;
-    border-radius: 20px;
-    padding: 1.15rem 1.35rem 1.25rem;
-    margin-bottom: 20px;
-    background: linear-gradient(160deg, rgba(191, 219, 254, 0.85), rgba(147, 197, 253, 0.65));
-    border: 1px solid rgba(59, 130, 246, 0.28);
-    box-shadow: 0 24px 46px rgba(30, 64, 175, 0.18);
-    color: #0f172a;
-    width: 100%;
-}
-
-.hero-info-card::after {
-    content: "";
-    position: absolute;
-    inset: 0;
-    border-radius: inherit;
-    background: linear-gradient(140deg, rgba(59, 130, 246, 0.12), rgba(29, 78, 216, 0.12));
-    opacity: 0;
-    transition: opacity 0.3s ease;
-    pointer-events: none;
-}
-
-.hero-info-card:hover::after {
-    opacity: 1;
-}
-
-.hero-info-card > * {
-    position: relative;
-    z-index: 1;
-}
-
-.hero-info-card h3 {
-    margin: 0 0 0.55rem;
-    font-size: 1.05rem;
-    color: #1d4ed8;
-}
-
-.hero-info-card p {
-    margin: 0;
-    font-size: 0.95rem;
-    line-height: 1.6;
-    color: rgba(15, 23, 42, 0.82);
-}
-
-.hero-info-steps {
-    margin: 0.55rem 0 0;
-    padding: 0;
-    list-style: none;
-    display: grid;
-    gap: 0.4rem;
-}
-
-.hero-info-steps li {
-    display: grid;
-    grid-template-columns: 32px minmax(0, 1fr);
-    gap: 0.5rem;
-    align-items: start;
-    font-size: 0.93rem;
-    color: rgba(15, 23, 42, 0.82);
-}
-
-.hero-info-steps .step-index {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    width: 32px;
-    height: 32px;
-    border-radius: 999px;
-    background: rgba(37, 99, 235, 0.18);
-    color: #1d4ed8;
-    font-weight: 600;
-    font-size: 0.85rem;
-    box-shadow: inset 0 0 0 1px rgba(37, 99, 235, 0.28);
-}
-
-.lifecycle-flow {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    gap: 0.6rem;
-    flex-wrap: wrap;
-    padding: 0.85rem 1rem;
-    margin-top: 0.85rem;
-    border-radius: 999px;
-    background: linear-gradient(120deg, rgba(191, 219, 254, 0.65), rgba(147, 197, 253, 0.35));
-    box-shadow: inset 0 0 0 1px rgba(59, 130, 246, 0.22);
-}
-
-.lifecycle-step {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.4rem;
-    padding: 0.35rem 0.75rem;
-    border-radius: 999px;
-    background: rgba(255, 255, 255, 0.85);
-    box-shadow: 0 12px 28px rgba(30, 64, 175, 0.18);
-    color: #1d4ed8;
-    font-weight: 600;
-    font-size: 0.95rem;
-}
-
-.lifecycle-step .lifecycle-icon {
-    font-size: 1.1rem;
-}
-
-.lifecycle-step .lifecycle-label {
-    color: #0f172a;
-    font-weight: 600;
-}
-
-.lifecycle-arrow {
-    font-size: 1.15rem;
-    color: rgba(15, 23, 42, 0.7);
-}
-
-.lifecycle-loop {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    width: 32px;
-    height: 32px;
-    border-radius: 50%;
-    background: rgba(30, 64, 175, 0.12);
-    color: #1d4ed8;
-    font-size: 1.1rem;
-    box-shadow: inset 0 0 0 1px rgba(37, 99, 235, 0.22);
-}
-
-.mailbox-summary-card .summary-list {
-    margin: 0.45rem 0 0;
-    padding-left: 1.1rem;
-    color: #1e293b;
-    font-size: 0.92rem;
-}
-
-.mailbox-summary-card .summary-list li {
-    margin-bottom: 0.25rem;
-}
-
-.mailbox-summary-card--empty {
-    text-align: center;
-    color: #475569;
-    font-style: italic;
-}
-
-.mailbox-summary-card--empty strong {
-    display: block;
-    margin-bottom: 0.35rem;
-}
-
-@media (max-width: 1100px) {
-    .mailbox-summary {
-        flex-direction: row;
-        flex-wrap: wrap;
-    }
-}
-
-[data-testid="stSidebar"] {
-    background: linear-gradient(185deg, rgba(226, 232, 240, 0.78), rgba(248, 250, 252, 0.96));
-    border-right: 1px solid rgba(148, 163, 184, 0.28);
-    color: #0f172a;
-}
-
-[data-testid="stSidebar"] > div:first-child {
-    padding: 2.6rem 1.6rem 2.1rem;
-}
-
-.sidebar-shell {
-    display: flex;
-    flex-direction: column;
-    gap: 2rem;
-}
-
-.sidebar-brand {
-    display: flex;
-    flex-direction: column;
-    gap: 0.4rem;
-}
-
-.sidebar-brand .sidebar-title {
-    margin: 0;
-    font-size: 1.18rem;
-    font-weight: 700;
-    color: #1d4ed8;
-}
-
-.sidebar-brand .sidebar-subtitle {
-    margin: 0;
-    color: rgba(15, 23, 42, 0.7);
-    line-height: 1.6;
-    font-size: 0.95rem;
-}
-
-.sidebar-nav {
-    display: flex;
-    flex-direction: column;
-    gap: 0.75rem;
-}
-
-.sidebar-nav [data-testid="stWidgetLabel"] > label {
-    font-size: 0.75rem;
-    letter-spacing: 0.08em;
-    text-transform: uppercase;
-    font-weight: 700;
-    color: rgba(15, 23, 42, 0.45);
-    margin-bottom: 0.2rem;
-}
-
-.sidebar-nav [role="radiogroup"] {
-    display: grid;
-    gap: 0.55rem;
-}
-
-.sidebar-nav [role="radiogroup"] label {
-    position: relative;
-    display: flex;
-    align-items: center;
-    gap: 0.65rem;
-    padding: 0.75rem 0.95rem;
-    border-radius: 16px;
-    border: 1px solid rgba(148, 163, 184, 0.4);
-    background: rgba(255, 255, 255, 0.82);
-    box-shadow: 0 14px 28px rgba(15, 23, 42, 0.12);
-    font-weight: 600;
-    color: #1e293b;
-    cursor: pointer;
-    transition: all 0.2s ease;
-}
-
-.sidebar-nav [role="radiogroup"] label:hover {
-    border-color: rgba(37, 99, 235, 0.45);
-    box-shadow: 0 20px 36px rgba(30, 64, 175, 0.18);
-    background: rgba(191, 219, 254, 0.6);
-}
-
-.sidebar-nav [role="radiogroup"] label input {
-    position: absolute;
-    inset: 0;
-    opacity: 0;
-    cursor: pointer;
-}
-
-.sidebar-nav [role="radiogroup"] label:has(input:checked) {
-    border-color: rgba(37, 99, 235, 0.55);
-    background: linear-gradient(150deg, rgba(37, 99, 235, 0.18), rgba(59, 130, 246, 0.32));
-    box-shadow: 0 24px 46px rgba(30, 64, 175, 0.22);
-}
-
-.sidebar-section-title {
-    font-size: 0.75rem;
-    letter-spacing: 0.08em;
-    text-transform: uppercase;
-    font-weight: 700;
-    color: rgba(15, 23, 42, 0.45);
-    margin-bottom: 0.3rem;
-}
-
-.sidebar-stage-card {
-    display: grid;
-    grid-template-columns: auto 1fr;
-    gap: 0.85rem;
-    padding: 0.95rem 1.05rem;
-    border-radius: 18px;
-    border: 1px solid rgba(148, 163, 184, 0.32);
-    background: rgba(255, 255, 255, 0.88);
-    box-shadow: 0 20px 42px rgba(15, 23, 42, 0.16);
-}
-
-.sidebar-stage-card__icon {
-    display: grid;
-    place-items: center;
-    width: 48px;
-    height: 48px;
-    border-radius: 16px;
-    background: rgba(59, 130, 246, 0.14);
-    color: #1d4ed8;
-    font-size: 1.3rem;
-    box-shadow: inset 0 0 0 1px rgba(59, 130, 246, 0.28);
-}
-
-.sidebar-stage-card__meta {
-    display: flex;
-    flex-direction: column;
-    gap: 0.2rem;
-}
-
-.sidebar-stage-card__eyebrow {
-    font-size: 0.7rem;
-    font-weight: 700;
-    letter-spacing: 0.1em;
-    text-transform: uppercase;
-    color: rgba(15, 23, 42, 0.55);
-}
-
-.sidebar-stage-card__title {
-    margin: 0;
-    font-size: 1.02rem;
-    font-weight: 700;
-    color: #0f172a;
-}
-
-.sidebar-stage-card__description {
-    margin: 0;
-    color: rgba(15, 23, 42, 0.7);
-    line-height: 1.55;
-    font-size: 0.9rem;
-}
-
-[data-testid="stSidebar"] .stToggle {
-    padding: 0.5rem 0.25rem 0.2rem;
-}
-
-[data-testid="stSidebar"] .stButton > button {
-    width: 100%;
-    border-radius: 14px;
-    border: none;
-    background: linear-gradient(140deg, #1d4ed8, #312e81);
-    color: #f8fafc;
-    font-weight: 600;
-    box-shadow: 0 18px 42px rgba(30, 64, 175, 0.28);
-}
-
-[data-testid="stSidebar"] .stButton > button:hover {
-    background: linear-gradient(140deg, #1e40af, #1d4ed8);
-    box-shadow: 0 22px 48px rgba(30, 64, 175, 0.32);
-}
-
-
-[data-testid="stSidebar"] .stButton > button:focus-visible {
-    outline: 3px solid rgba(59, 130, 246, 0.55);
-    outline-offset: 2px;
-}
-
-.metric-highlight {
-    display: flex;
-    align-items: center;
-    gap: 0.75rem;
-    padding: 0.85rem 1.1rem;
-    border-radius: 16px;
-    background: rgba(52, 211, 153, 0.14);
-    border: 1px solid rgba(52, 211, 153, 0.28);
-    color: #047857;
-    font-weight: 600;
-    margin: 1rem 0;
-}
-
-.metric-highlight svg {
-    width: 28px;
-    height: 28px;
-}
-
-.pill-group {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 0.6rem;
-    margin: 1rem 0;
-}
-
-.pill-group .pill {
-    padding: 0.45rem 0.9rem;
-    border-radius: 999px;
-    background: rgba(59, 130, 246, 0.12);
-    color: #1d4ed8;
-    font-size: 0.85rem;
-    font-weight: 600;
-}
-</style>
-"""
-
 st.markdown(APP_THEME_CSS, unsafe_allow_html=True)
-
-STAGE_TEMPLATE_CSS = """
-<style>
-:root {
-    --stage-card-radius: 18px;
-    --stage-card-border: rgba(15, 23, 42, 0.08);
-    --stage-card-shadow: 0 18px 40px rgba(15, 23, 42, 0.10);
-}
-
-.stage-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
-    gap: 1.2rem;
-    margin-top: 1.2rem;
-}
-
-.stage-progress-grid {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 0.6rem;
-    margin: 1rem 0 0;
-    padding: 0.5rem 0.75rem;
-    border-radius: 16px;
-    background: rgba(15, 23, 42, 0.03);
-    border: 1px solid rgba(148, 163, 184, 0.24);
-    overflow-x: auto;
-}
-
-.stage-progress-grid::-webkit-scrollbar {
-    height: 6px;
-}
-
-.stage-progress-grid::-webkit-scrollbar-thumb {
-    background: rgba(100, 116, 139, 0.35);
-    border-radius: 999px;
-}
-
-.stage-card {
-    background: linear-gradient(145deg, rgba(255, 255, 255, 0.96), rgba(241, 245, 255, 0.9));
-    border-radius: var(--stage-card-radius);
-    border: 1px solid var(--stage-card-border);
-    box-shadow: var(--stage-card-shadow);
-    padding: 1.15rem 1.2rem;
-    transition: transform 0.2s ease, box-shadow 0.2s ease, border-color 0.2s ease;
-    color: #0f172a;
-    position: relative;
-    overflow: hidden;
-    display: block;
-    text-decoration: none;
-}
-
-.stage-card:link,
-.stage-card:visited {
-    color: #0f172a;
-    text-decoration: none;
-}
-
-.stage-card::after {
-    content: "";
-    position: absolute;
-    inset: 0;
-    background: radial-gradient(circle at top right, rgba(74, 108, 247, 0.18), transparent 55%);
-    opacity: 0;
-    transition: opacity 0.2s ease;
-}
-
-.stage-card:hover {
-    transform: translateY(-3px);
-    box-shadow: 0 22px 50px rgba(15, 23, 42, 0.14);
-}
-
-.stage-card:hover::after {
-    opacity: 1;
-}
-
-.stage-card:focus-visible {
-    outline: 3px solid rgba(74, 108, 247, 0.45);
-    outline-offset: 3px;
-}
-
-.stage-card .stage-icon {
-    font-size: 1.85rem;
-    line-height: 1;
-    margin-bottom: 0.35rem;
-}
-
-.stage-card .stage-title {
-    font-size: 1.05rem;
-    font-weight: 600;
-}
-
-.stage-card .stage-summary {
-    margin-top: 0.55rem;
-    font-size: 0.92rem;
-    line-height: 1.5;
-    color: #334155;
-}
-
-.stage-card.active {
-    border-color: rgba(74, 108, 247, 0.55);
-    background: linear-gradient(155deg, rgba(74, 108, 247, 0.18), rgba(241, 245, 255, 0.96));
-}
-
-.stage-card.complete {
-    border-color: rgba(52, 199, 123, 0.55);
-    background: linear-gradient(155deg, rgba(52, 199, 123, 0.18), rgba(240, 253, 244, 0.95));
-}
-
-.stage-progress-grid .stage-card {
-    min-height: auto;
-    padding: 0.45rem 0.75rem;
-    border-radius: 999px;
-    box-shadow: none;
-    background: rgba(255, 255, 255, 0.94);
-    display: inline-flex;
-    align-items: center;
-    gap: 0.45rem;
-    text-align: left;
-    cursor: pointer;
-}
-
-.stage-progress-grid .stage-card.complete {
-    background: linear-gradient(145deg, rgba(52, 199, 123, 0.16), rgba(240, 253, 244, 0.92));
-    border-color: rgba(52, 199, 123, 0.45);
-}
-
-.stage-progress-grid .stage-card.active {
-    background: linear-gradient(145deg, rgba(74, 108, 247, 0.16), rgba(241, 245, 255, 0.95));
-    border-color: rgba(74, 108, 247, 0.45);
-    box-shadow: 0 10px 24px rgba(37, 99, 235, 0.15);
-}
-
-.stage-progress-grid .stage-icon {
-    font-size: 1.15rem;
-    margin-bottom: 0;
-}
-
-.stage-progress-grid .stage-summary {
-    display: none;
-}
-
-.stage-progress-grid .stage-title {
-    font-size: 0.9rem;
-    font-weight: 600;
-}
-
-.stage-nav-buttons {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 0.75rem;
-    margin-top: 1.75rem;
-}
-
-.stage-nav-buttons .stButton>button {
-    flex: 1;
-    min-width: 200px;
-    border-radius: 999px;
-    font-weight: 600;
-    padding: 0.65rem 1.1rem;
-    box-shadow: 0 10px 24px rgba(15, 23, 42, 0.08);
-}
-
-.stage-nav-buttons .stButton>button:hover {
-    box-shadow: 0 16px 32px rgba(15, 23, 42, 0.12);
-}
-
-.stage-nav-buttons .stButton>button:focus {
-    outline: none;
-    box-shadow: 0 0 0 3px rgba(74, 108, 247, 0.35);
-}
-
-.ai-quote-box {
-    margin: 1.35rem 0;
-    padding: 1.4rem 1.6rem;
-    border-radius: 20px;
-    border: 1px solid rgba(37, 99, 235, 0.28);
-    background: linear-gradient(140deg, rgba(219, 234, 254, 0.82), rgba(191, 219, 254, 0.45));
-    color: #0f172a;
-    display: grid;
-    grid-template-columns: auto 1fr;
-    gap: 1rem;
-    align-items: flex-start;
-    box-shadow: 0 18px 42px rgba(30, 64, 175, 0.16);
-    position: relative;
-    overflow: hidden;
-}
-
-.ai-quote-box::before {
-    content: "";
-    position: absolute;
-    inset: 0;
-    border-radius: inherit;
-    background: linear-gradient(160deg, rgba(37, 99, 235, 0.18), transparent 62%);
-    pointer-events: none;
-}
-
-.ai-quote-box__icon {
-    font-size: 1.8rem;
-    line-height: 1;
-    color: #1d4ed8;
-    position: relative;
-    z-index: 1;
-    margin-top: 0.2rem;
-}
-
-.ai-quote-box__content {
-    position: relative;
-    z-index: 1;
-    display: flex;
-    flex-direction: column;
-    gap: 0.45rem;
-}
-
-.ai-quote-box__source {
-    font-size: 0.82rem;
-    font-weight: 600;
-    letter-spacing: 0.12em;
-    text-transform: uppercase;
-    color: rgba(30, 64, 175, 0.75);
-}
-
-.ai-quote-box p {
-    margin: 0;
-    font-size: 1rem;
-    line-height: 1.6;
-    color: #0f172a;
-}
-</style>
-"""
-
-EMAIL_INBOX_TABLE_CSS = """
-<style>
-.email-inbox-wrapper {
-    border: 1px solid rgba(15, 23, 42, 0.12);
-    border-radius: 14px;
-    overflow: hidden;
-    box-shadow: 0 18px 40px rgba(15, 23, 42, 0.12);
-    background: linear-gradient(145deg, rgba(255, 255, 255, 0.96), rgba(241, 245, 255, 0.92));
-    margin: 0.75rem 0 1.1rem;
-}
-
-.email-inbox-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 0.65rem 1.1rem;
-    background: linear-gradient(120deg, rgba(30, 64, 175, 0.9), rgba(59, 130, 246, 0.85));
-    color: #f8fafc;
-    font-weight: 600;
-    font-size: 0.94rem;
-}
-
-.email-inbox-table {
-    width: 100%;
-    border-collapse: collapse;
-    font-size: 0.9rem;
-    color: #0f172a;
-}
-
-.email-inbox-table thead {
-    background: rgba(226, 232, 240, 0.75);
-    text-transform: uppercase;
-    letter-spacing: 0.04em;
-    font-size: 0.76rem;
-    color: #1e293b;
-}
-
-.email-inbox-table th,
-.email-inbox-table td {
-    padding: 0.75rem 1rem;
-    border-bottom: 1px solid rgba(148, 163, 184, 0.35);
-    vertical-align: top;
-}
-
-.email-inbox-table tbody tr:hover {
-    background: rgba(191, 219, 254, 0.45);
-}
-
-.email-inbox-table tbody tr:nth-child(even) {
-    background: rgba(248, 250, 252, 0.85);
-}
-
-.email-inbox-empty {
-    padding: 1rem 1.25rem;
-    color: #475569;
-    font-style: italic;
-}
-</style>
-"""
-
-LIFECYCLE_CYCLE_CSS = """
-<style>
-.lifecycle-cycle {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-wrap: wrap;
-    gap: 0.9rem;
-    margin: 1.2rem 0 1.6rem;
-}
-
-.lifecycle-cycle .cycle-node {
-    width: 118px;
-    height: 118px;
-    border-radius: 50%;
-    background: linear-gradient(160deg, rgba(59, 130, 246, 0.15), rgba(59, 130, 246, 0.05));
-    border: 2px solid rgba(37, 99, 235, 0.35);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-    padding: 0.65rem;
-    color: #1e3a8a;
-    font-weight: 600;
-    box-shadow: inset 0 0 12px rgba(30, 64, 175, 0.18), 0 18px 34px rgba(15, 23, 42, 0.12);
-    transition: transform 0.2s ease, box-shadow 0.2s ease;
-}
-
-.lifecycle-cycle .cycle-node .cycle-icon {
-    font-size: 1.8rem;
-    margin-bottom: 0.35rem;
-}
-
-.lifecycle-cycle .cycle-node-active {
-    background: linear-gradient(160deg, rgba(52, 211, 153, 0.25), rgba(52, 211, 153, 0.1));
-    border-color: rgba(5, 150, 105, 0.55);
-    color: #065f46;
-    transform: scale(1.04);
-    box-shadow: inset 0 0 16px rgba(16, 185, 129, 0.25), 0 20px 40px rgba(13, 148, 136, 0.18);
-}
-
-.lifecycle-cycle .cycle-arrow {
-    font-size: 1.75rem;
-    color: rgba(30, 64, 175, 0.65);
-}
-
-.lifecycle-cycle .cycle-loop {
-    width: 64px;
-    height: 64px;
-    border-radius: 50%;
-    border: 2px dashed rgba(30, 64, 175, 0.35);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 1.65rem;
-    color: rgba(30, 64, 175, 0.7);
-    box-shadow: inset 0 0 12px rgba(37, 99, 235, 0.15);
-}
-</style>
-"""
-
 st.markdown(STAGE_TEMPLATE_CSS, unsafe_allow_html=True)
-
-
-@contextmanager
-def section_surface(class_name: str = ""):
-    block = st.container()
-    marker_id = f"section-surface-{uuid4().hex}"
-    classes = "section-surface" + (f" {class_name}" if class_name else "")
-    marker = block.empty()
-    marker.markdown(f"<div id='{marker_id}' class='{classes}'></div>", unsafe_allow_html=True)
-
-    applied_classes = class_name.split() if class_name else []
-    class_script = "".join(f"block.classList.add('{cls}');" for cls in applied_classes)
-    components.html(
-        f"""
-<script>
-(function() {{
-  let attempts = 0;
-  const MAX_ATTEMPTS = 40;
-  const attach = () => {{
-    const rootDoc = window.parent?.document;
-    if (!rootDoc) {{
-      if (attempts++ < MAX_ATTEMPTS) {{
-        setTimeout(attach, 50);
-      }}
-      return;
-    }}
-    const marker = rootDoc.getElementById('{marker_id}');
-    if (!marker) {{
-      if (attempts++ < MAX_ATTEMPTS) {{
-        setTimeout(attach, 50);
-      }}
-      return;
-    }}
-    const elementContainer = marker.closest('[data-testid="stElementContainer"]');
-    const wrapper = elementContainer?.parentElement?.nextElementSibling;
-    const block = wrapper?.querySelector('[data-testid="stVerticalBlock"]') ?? marker.closest('[data-testid="stVerticalBlock"]');
-    if (!block) {{
-      if (attempts++ < MAX_ATTEMPTS) {{
-        setTimeout(attach, 50);
-        return;
-      }}
-      marker.classList.remove('section-surface');
-      marker.style.display = 'block';
-      return;
-    }}
-    block.classList.add('section-surface-block');
-    {class_script}
-    if (elementContainer) {{
-      elementContainer.remove();
-    }} else {{
-      marker.remove();
-    }}
-  }};
-  attach();
-}})();
-</script>
-""",
-        height=0,
-    )
-
-    with block:
-        yield
-
-
-def render_nerd_mode_toggle(
-    *,
-    key: str,
-    title: str = "Nerd Mode",
-    description: Optional[str] = None,
-    icon: str = "🧠",
-    default: Optional[bool] = None,
-) -> bool:
-    """Render a styled Nerd Mode toggle and return its value."""
-
-    existing = st.session_state.get(key, None)
-    initial_value = existing if existing is not None else default
-
-    container = st.container()
-    with container:
-        st.markdown("<div class='nerd-toggle-card'>", unsafe_allow_html=True)
-        info_col, toggle_col = st.columns([4, 1])
-        with info_col:
-            st.markdown(
-                f"<div class='nerd-toggle-title'>{html.escape(icon)} {html.escape(title)}</div>",
-                unsafe_allow_html=True,
-            )
-            if description:
-                st.markdown(
-                    f"<div class='nerd-toggle-description'>{html.escape(description)}</div>",
-                    unsafe_allow_html=True,
-                )
-        with toggle_col:
-            value = st.toggle(
-                title,
-                key=key,
-                value=bool(initial_value) if initial_value is not None else False,
-                label_visibility="collapsed",
-            )
-        st.markdown("</div>", unsafe_allow_html=True)
-
-    return value
-
-
-def _stage_card_html(
-    stage: StageMeta,
-    status_class: str,
-    *,
-    show_description: bool,
-    clickable: bool,
-) -> str:
-    body_text = stage.description if show_description else stage.summary
-    classes = "stage-card" + (f" {status_class}" if status_class else "")
-    tag = "a" if clickable else "div"
-    href_attr = ""
-    if clickable:
-        params = {k: list(st.query_params.get_all(k)) for k in st.query_params}
-        params["stage"] = [stage.key]
-        href_attr = f" href=\"?{urlencode(params, doseq=True)}\" target=\"_self\""
-    aria_attr = " aria-current=\"step\"" if status_class == "active" else ""
-    return (
-        f"<{tag} class=\"{classes}\"{href_attr}{aria_attr}>"
-        "<div class=\"stage-icon\">{icon}</div>"
-        "<div class=\"stage-title\">{title}</div>"
-        "<div class=\"stage-summary\">{summary}</div>"
-        f"</{tag}>"
-    ).format(
-        icon=html.escape(stage.icon),
-        title=html.escape(stage.title),
-        summary=html.escape(body_text),
-    )
-
-
-def render_stage_cards(
-    active_stage: str,
-    *,
-    variant: str = "grid",
-    stages: Optional[List[StageMeta]] = None,
-):
-    """Render the reusable stage template as a grid or compact progress bar."""
-
-    wrapper_class = "stage-grid" if variant == "grid" else "stage-progress-grid"
-    show_description = variant == "grid"
-    active_index = STAGE_INDEX.get(active_stage, 0)
-    stage_list = stages if stages is not None else STAGES
-    clickable = variant != "grid"
-    cards: List[str] = []
-    for stage in stage_list:
-        idx = STAGE_INDEX[stage.key]
-        if idx < active_index:
-            status_class = "complete"
-        elif idx == active_index:
-            status_class = "active"
-        else:
-            status_class = ""
-        cards.append(
-            _stage_card_html(
-                stage,
-                status_class=status_class,
-                show_description=show_description,
-                clickable=clickable,
-            )
-        )
-
-    st.markdown(
-        f"<div class=\"{wrapper_class}\">{''.join(cards)}</div>",
-        unsafe_allow_html=True,
-    )
-
-
-def set_active_stage(stage_key: str):
-    if stage_key not in STAGE_BY_KEY:
-        return
-    st.session_state["active_stage"] = stage_key
-    if st.query_params.get_all("stage") != [stage_key]:
-        st.query_params["stage"] = stage_key
-
-
-def render_stage_navigation_controls(active_stage: str):
-    idx = STAGE_INDEX.get(active_stage, 0)
-    prev_stage = STAGES[idx - 1] if idx > 0 else None
-    next_stage = STAGES[idx + 1] if idx < len(STAGES) - 1 else None
-
-    st.markdown("<div class='stage-nav-buttons'>", unsafe_allow_html=True)
-    col_prev, col_status, col_next = st.columns([1.2, 0.9, 1.2])
-
-    with col_prev:
-        if prev_stage:
-            st.button(
-                f"⬅️ Back to {prev_stage.title}",
-                key=f"nav_back_{active_stage}",
-                on_click=set_active_stage,
-                args=(prev_stage.key,),
-                width="stretch",
-            )
-        else:
-            st.write(" ")
-
-    with col_status:
-        if active_stage != "intro":
-            stage = STAGE_BY_KEY[active_stage]
-            st.markdown(
-                """
-                <div style="text-align:center; padding:0.65rem 0.75rem; background: rgba(15, 23, 42, 0.03); border-radius: 14px; font-weight: 600; color: #475569;">
-                    Currently on <span style="color:#1d4ed8;">{icon} {title}</span>
-                </div>
-                """.format(icon=html.escape(stage.icon), title=html.escape(stage.title)),
-                unsafe_allow_html=True,
-            )
-        else:
-            st.write(" ")
-
-    with col_next:
-        if next_stage and active_stage != "intro":
-            st.button(
-                f"Next • {next_stage.title} ➡️",
-                key=f"nav_next_{active_stage}",
-                on_click=set_active_stage,
-                args=(next_stage.key,),
-                width="stretch",
-                type="primary",
-            )
-        else:
-            st.write(" ")
-
-    st.markdown("</div>", unsafe_allow_html=True)
-
-
-def render_email_inbox_table(
-    df: pd.DataFrame,
-    *,
-    title: str,
-    subtitle: Optional[str] = None,
-    columns: Optional[List[str]] = None,
-) -> None:
-    st.markdown(EMAIL_INBOX_TABLE_CSS, unsafe_allow_html=True)
-
-    if df is None or df.empty:
-        header_html = (
-            f"<div class='email-inbox-header'><span>{html.escape(title)}</span><span>0 messages</span></div>"
-        )
-        body_html = html.escape(subtitle) if subtitle else "Inbox is empty."
-        st.markdown(
-            f"<div class='email-inbox-wrapper'>{header_html}<div class='email-inbox-empty'>{body_html}</div></div>",
-            unsafe_allow_html=True,
-        )
-        return
-
-    display_df = df.copy()
-    if columns:
-        existing_cols = [col for col in columns if col in display_df.columns]
-        if existing_cols:
-            display_df = display_df[existing_cols]
-
-    if "title" in display_df.columns:
-        display_df.rename(columns={"title": "Subject"}, inplace=True)
-
-    if "body" in df.columns:
-        preview_series = df["body"].fillna("").apply(
-            lambda text: text if len(text) <= 90 else text[:87].rstrip() + "…"
-        )
-        insert_position = len(display_df.columns)
-        if "Subject" in display_df.columns:
-            insert_position = list(display_df.columns).index("Subject") + 1
-        display_df.insert(insert_position, "Preview", preview_series)
-        if "body" in display_df.columns:
-            display_df.drop(columns=["body"], inplace=True)
-
-    header_right = f"{len(df)} message{'s' if len(df) != 1 else ''}"
-    header_html = f"<div class='email-inbox-header'><span>{html.escape(title)}</span><span>{html.escape(header_right)}</span></div>"
-    subtitle_html = (
-        f"<div style='padding:0.45rem 1.1rem; color:#475569;'>{html.escape(subtitle)}</div>"
-        if subtitle
-        else ""
-    )
-    table_html = display_df.to_html(classes="email-inbox-table", index=False, escape=False)
-    st.markdown(
-        f"<div class='email-inbox-wrapper'>{header_html}{subtitle_html}{table_html}</div>",
-        unsafe_allow_html=True,
-    )
-
-
-def render_mailbox_summary(df: pd.DataFrame, mailbox_title: str) -> None:
-    st.markdown(f"#### Quick insights — {mailbox_title}")
-
-    if df is None or df.empty:
-        st.markdown(
-            f"""
-            <div class="mailbox-summary">
-                <div class="mailbox-summary-card mailbox-summary-card--empty">
-                    <strong>No messages yet in {html.escape(mailbox_title)}</strong>
-                    As soon as emails appear in this mailbox, a snapshot of activity will show up here.
-                </div>
-            </div>
-            """,
-            unsafe_allow_html=True,
-        )
-        return
-
-    total = len(df)
-    avg_spam = float(df["p_spam"].mean()) if "p_spam" in df.columns else 0.0
-
-    prediction_rows = ""
-    if "pred" in df.columns and not df["pred"].isna().all():
-        counts = df["pred"].value_counts().sort_values(ascending=False)
-        prediction_rows = "".join(
-            f"<div class='summary-row'><span>{html.escape(str(label).title())}</span><span>{count} ({count / total:.0%})</span></div>"
-            for label, count in counts.items()
-        )
-    if not prediction_rows:
-        prediction_rows = "<div class='summary-row'>No predictions recorded yet.</div>"
-
-    subjects: List[str] = []
-    if "title" in df.columns:
-        subjects = []
-        for subject in df["title"].fillna(""):
-            subject_text = str(subject).strip()
-            if subject_text:
-                subjects.append(html.escape(subject_text))
-            if len(subjects) == 3:
-                break
-    subject_items = "".join(f"<li>{item}</li>" for item in subjects)
-    if not subject_items:
-        subject_items = "<li>No subject lines available yet.</li>"
-
-    threshold = float(ss.get("threshold", 0.5))
-
-    summary_html = f"""
-    <div class="mailbox-summary">
-        <div class="mailbox-summary-card">
-            <div class="summary-title">Messages</div>
-            <div class="summary-value">{total}</div>
-            <div class="summary-hint">Avg. spam score: {avg_spam:.0%}</div>
-        </div>
-        <div class="mailbox-summary-card">
-            <div class="summary-title">Prediction mix</div>
-            <div class="summary-rows">{prediction_rows}</div>
-        </div>
-        <div class="mailbox-summary-card">
-            <div class="summary-title">Latest subjects</div>
-            <ul class="summary-list">{subject_items}</ul>
-            <div class="summary-hint">Routing threshold: {threshold:.2f}</div>
-        </div>
-    </div>
-    """
-    st.markdown(summary_html, unsafe_allow_html=True)
-
-
-def render_mailbox_panel(
-    records: Optional[List[Dict[str, Any]]],
-    *,
-    mailbox_title: str,
-    filled_subtitle: str,
-    empty_subtitle: str,
-) -> None:
-    df_raw = pd.DataFrame(records) if records else pd.DataFrame()
-    table_col, summary_col = st.columns((2.8, 1.2))
-
-    with table_col:
-        if df_raw.empty:
-            render_email_inbox_table(pd.DataFrame(), title=mailbox_title, subtitle=empty_subtitle)
-        else:
-            display_df = df_raw.rename(columns={"pred": "Prediction", "p_spam": "P(spam)"})
-            render_email_inbox_table(
-                display_df,
-                title=mailbox_title,
-                subtitle=filled_subtitle,
-            )
-
-    with summary_col:
-        render_mailbox_summary(df_raw, mailbox_title)
-
-
-def render_lifecycle_cycle(active_stage: str) -> None:
-    st.markdown(LIFECYCLE_CYCLE_CSS, unsafe_allow_html=True)
-
-    ordered_keys = ["data", "train", "evaluate", "classify"]
-    parts: List[str] = []
-    for idx, key in enumerate(ordered_keys):
-        stage = STAGE_BY_KEY[key]
-        active_cls = " cycle-node-active" if key == active_stage else ""
-        parts.append(
-            """
-            <div class="cycle-node{active_cls}">
-                <span class="cycle-icon">{icon}</span>
-                <span class="cycle-label">{title}</span>
-            </div>
-            """.format(
-                active_cls=active_cls,
-                icon=html.escape(stage.icon),
-                title=html.escape(stage.title),
-            )
-        )
-        if idx < len(ordered_keys) - 1:
-            parts.append("<div class='cycle-arrow'>➝</div>")
-
-    parts.append("<div class='cycle-loop' title='Back to the start'>↺</div>")
-    st.markdown(f"<div class='lifecycle-cycle'>{''.join(parts)}</div>", unsafe_allow_html=True)
-
-CLASSES = ["spam", "safe"]
-AUTONOMY_LEVELS = [
-    "Moderate autonomy (recommendation)",
-    "High autonomy (auto-route)",
-]
-
-@dataclass(frozen=True)
-class StageMeta:
-    key: str
-    title: str
-    icon: str
-    summary: str
-    description: str
-
-
-STAGES: List[StageMeta] = [
-    StageMeta("intro", "Welcome", "🚀", "Kickoff", "Meet the mission and trigger the guided build."),
-    StageMeta(
-        "overview",
-        "Start your machine",
-        "🧭",
-        "See the journey",
-        "Tour the steps, set Nerd Mode, and align on what you'll do.",
-    ),
-    StageMeta("data", "Prepare Data", "📊", "Curate examples", "Inspect labeled emails and get the dataset ready for learning."),
-    StageMeta("train", "Train", "🧠", "Teach the model", "Configure the split and teach the model on your dataset."),
-    StageMeta("evaluate", "Evaluate", "🧪", "Check results", "Check metrics, inspect the confusion matrix, and stress-test."),
-    StageMeta("classify", "Use", "📬", "Route emails", "Route new messages, correct predictions, and adapt."),
-    StageMeta("model_card", "Model Card", "📄", "Document system", "Summarize performance, intended use, and governance."),
-]
-
-STAGE_INDEX = {stage.key: idx for idx, stage in enumerate(STAGES)}
-STAGE_BY_KEY = {stage.key: stage for stage in STAGES}
-
-
-DATASET_SUSPICIOUS_TLDS = [
-    ".ru",
-    ".xyz",
-    ".top",
-    ".biz",
-    ".win",
-    ".loan",
-    ".live",
-    ".icu",
-    ".cn",
-    ".gq",
-]
-
-DATASET_LEGIT_DOMAINS = [
-    "intranet.corp",
-    "sharepoint.corp",
-    "confluence.corp",
-    "workday.corp",
-    "hr.corp",
-    "vpn.corp",
-    "it.corp",
-]
-
-BRANDS = ["DocuSign", "Office 365", "OneDrive", "Zoom", "Adobe", "Okta", "Teams"]
-
-COURIERS = ["DHL", "UPS", "FedEx", "PostNL", "PostNord"]
-
-URGENCY = ["URGENT", "ACTION REQUIRED", "FINAL NOTICE", "IMMEDIATE", "24H", "TODAY"]
-
-def _rand_amount(rng: random.Random) -> str:
-    euros = rng.choice(
-        [
-            18,
-            44.09,
-            67.45,
-            73.33,
-            89.63,
-            120.07,
-            190.05,
-            250.52,
-            318.77,
-            540.54,
-            725.25,
-            930.39,
-            1299.99,
-            1850.75,
-            2875.10,
-            4140.00,
-            7420.88,
-            9999.01,
-        ]
-    )
-    return f"€{euros}"
-
-
-def _maybe_caps(rng: random.Random, text: str, prob: float = 0.25) -> str:
-    if rng.random() > prob:
-        return text
-    words = text.split()
-    if not words:
-        return text
-    idx = rng.randrange(len(words))
-    words[idx] = words[idx].upper()
-    if rng.random() < 0.15 and len(words) > 2:
-        j = rng.randrange(len(words))
-        if j != idx:
-            words[j] = words[j].upper()
-    return " ".join(words)
-
-
-def _spam_link(rng: random.Random) -> str:
-    name = "".join(rng.choices(string.ascii_lowercase, k=rng.randint(6, 12)))
-    tld = rng.choice(DATASET_SUSPICIOUS_TLDS)
-    path = "".join(rng.choices(string.ascii_lowercase, k=rng.randint(4, 10)))
-    scheme = rng.choice(["http://", "https://"])
-    return f"{scheme}{name}{tld}/{path}"
-
-
-def _safe_link(rng: random.Random) -> str:
-    host = rng.choice(DATASET_LEGIT_DOMAINS)
-    path = rng.choice(["/docs", "/policies", "/training", "/tickets", "/files", "/benefits"])
-    return f"https://{host}{path}"
-
-
-def _maybe_links(rng: random.Random, k: int, suspicious: bool = True) -> str:
-    make = _spam_link if suspicious else _safe_link
-    return " ".join(make(rng) for _ in range(k))
-
-
-def _maybe_attachment(rng: random.Random, kind: str = "spam") -> str:
-    if kind == "spam":
-        return rng.choice(
-            [
-                "Open the attached HTML and sign in.",
-                "Download the attached ZIP and run the installer.",
-                "Enable macros in the attached XLSM to proceed.",
-                "Install the attached EXE to restore access.",
-                "Open the PDF and confirm your credentials.",
-            ]
-        )
-    return rng.choice(
-        [
-            "Please see the attached PDF; no further action required.",
-            "Slides attached for review.",
-            "Agenda attached; join via Teams.",
-            "Invoice PDF attached; PO noted.",
-            "Itinerary PDF attached; MFA required to view.",
-        ]
-    )
-
-
-def _spam_title_body(rng: random.Random) -> tuple[str, str]:
-    archetype = rng.choice(
-        [
-            "payroll_hold",
-            "account_reset",
-            "delivery_fee",
-            "invoice_wire",
-            "prize_lottery",
-            "crypto_double",
-            "bonus_verify",
-            "docu_phish",
-            "refund_now",
-            "mfa_disable",
-            "tax_rebate",
-        ]
-    )
-    amt = _rand_amount(rng)
-    brand = rng.choice(BRANDS)
-    courier = rng.choice(COURIERS)
-    urgency = rng.choice(URGENCY)
-
-    if archetype == "payroll_hold":
-        title = f"{urgency}: Verify your payroll to release deposit"
-        body = (
-            "Your salary is on hold. Confirm bank details at "
-            f"{_spam_link(rng)} within 30 minutes to avoid delay. {_maybe_attachment(rng, 'spam')}"
-        )
-    elif archetype == "account_reset":
-        title = f"Password will expire — {urgency.lower()}"
-        body = (
-            f"Reset your password here: {_spam_link(rng)}. "
-            "Failure to act may lock your account. Enter your email and password to confirm."
-        )
-    elif archetype == "delivery_fee":
-        title = f"{courier} delivery notice: customs fee required"
-        body = f"Your parcel is pending. Pay a small fee ({amt}) to schedule a new delivery slot at {_spam_link(rng)}."
-    elif archetype == "invoice_wire":
-        title = "Payment overdue — settle immediately"
-        body = f"Service interruption imminent. Wire {amt} to the account in the attachment today. {_maybe_attachment(rng, 'spam')}"
-    elif archetype == "prize_lottery":
-        title = "WIN a FREE gift — final eligibility"
-        body = f"Congratulations! Complete the short form at {_spam_link(rng)} to claim your reward. Offer expires in 2 hours."
-    elif archetype == "crypto_double":
-        title = "Crypto opportunity: double your balance"
-        body = f"Transfer funds and we’ll return 2× within 24h. Start at {_spam_link(rng)}. Trusted by executives."
-    elif archetype == "bonus_verify":
-        title = "HR update: bonus identity check"
-        body = f"Confirm your identity using your national ID and card CVV at {_spam_link(rng)} to receive your bonus."
-    elif archetype == "docu_phish":
-        title = f"{brand}: You received a secure document"
-        body = f"Open the external portal at {_spam_link(rng)} and log in with your email password to view the document."
-    elif archetype == "refund_now":
-        title = "Refund available — action needed"
-        body = f"We owe you {amt}. Submit IBAN and CVV at {_spam_link(rng)} to receive payment now."
-    elif archetype == "mfa_disable":
-        title = "Two-factor disabled — reactivate now"
-        body = f"We turned off MFA on your account. Re-enable at {_spam_link(rng)} (login required)."
-    elif archetype == "tax_rebate":
-        title = "Tax rebate waiting — confirm identity"
-        body = f"Claim your rebate by verifying bank access at {_spam_link(rng)}. {_maybe_attachment(rng, 'spam')}"
-    else:
-        title = "Important notice"
-        body = f"Complete the verification at {_spam_link(rng)}."
-
-    title = _maybe_caps(rng, title, prob=0.35)
-    if rng.random() < 0.25:
-        title += " !!"
-    if rng.random() < 0.35:
-        body += " Act NOW."
-    if rng.random() < 0.3:
-        body += f" More info: {_spam_link(rng)}"
-
-    return title, body
-
-
-def _safe_title_body(rng: random.Random) -> tuple[str, str]:
-    archetype = rng.choice(
-        [
-            "meeting",
-            "policy_update",
-            "hr_workday",
-            "it_maintenance",
-            "invoice_legit",
-            "travel",
-            "training",
-            "vendor_ap",
-            "security_advice",
-            "delivery_tracking",
-        ]
-    )
-    brand = rng.choice(BRANDS)
-    courier = rng.choice(COURIERS)
-
-    if archetype == "meeting":
-        title = "Team meeting moved to 14:00"
-        body = "Join via the usual Teams link. Agenda: KPIs, risk register, roadmap. " + _maybe_attachment(rng, "safe")
-    elif archetype == "policy_update":
-        title = "Policy update: remote work guidelines"
-        body = "Please review the updated policy on the intranet. " + _maybe_links(rng, 1, suspicious=False)
-    elif archetype == "hr_workday":
-        title = "Workday: benefits enrollment opens"
-        body = "Make your selections in Workday before month end. No personal info by email."
-    elif archetype == "it_maintenance":
-        title = "IT maintenance window"
-        body = "Patching on Saturday 22:00–23:30 CET. Expect brief reboots; no action required."
-    elif archetype == "invoice_legit":
-        title = "Invoice attached — Accounts Payable"
-        body = "Please find the invoice PDF attached. PO is referenced; no payment info requested. " + _maybe_attachment(rng, "safe")
-    elif archetype == "travel":
-        title = "Travel itinerary update"
-        body = "Platform change noted. PDF itinerary attached; bookings remain via the internal tool."
-    elif archetype == "training":
-        title = "Mandatory training assigned"
-        body = "Complete the e-learning module by next Friday. Materials on the LMS. " + _maybe_links(rng, 1, suspicious=False)
-    elif archetype == "vendor_ap":
-        title = "AP reminder — PO mismatch"
-        body = "Please correct the PO reference in the invoice metadata on SharePoint. " + _maybe_links(rng, 1, suspicious=False)
-    elif archetype == "security_advice":
-        title = f"Security advisory — {brand} tips"
-        body = "Review the guidance on the internal portal. No external logins required. " + _maybe_links(rng, 1, suspicious=False)
-    elif archetype == "delivery_tracking":
-        title = f"{courier} tracking: package out for delivery"
-        body = f"Your parcel is scheduled today. Track via the official {courier} site with your tracking ID (no payment)."
-    else:
-        title = "Update posted"
-        body = "Details are available on the intranet."
-
-    title = _maybe_caps(rng, title, prob=0.05)
-    if rng.random() < 0.1:
-        body += " Thanks."
-    if rng.random() < 0.2:
-        body += " Reference: " + _maybe_links(rng, 1, suspicious=False)
-
-    return title, body
-
-
-class DatasetConfig(TypedDict, total=False):
-    seed: int
-    n_total: int
-    spam_ratio: float
-    susp_link_level: str
-    susp_tld_level: str
-    caps_intensity: str
-    money_urgency: str
-    attachments_mix: Dict[str, float]
-    edge_cases: int
-    label_noise_pct: float
-    poison_demo: bool
-
-
-ATTACHMENT_TYPES = ["html", "zip", "xlsm", "exe", "pdf"]
-DEFAULT_ATTACHMENT_MIX: Dict[str, float] = {"html": 0.15, "zip": 0.15, "xlsm": 0.1, "exe": 0.1, "pdf": 0.5}
-ATTACHMENT_MIX_PRESETS: Dict[str, Dict[str, float]] = {
-    "Mostly PDF": {"html": 0.05, "zip": 0.05, "xlsm": 0.05, "exe": 0.05, "pdf": 0.80},
-    "Balanced": DEFAULT_ATTACHMENT_MIX.copy(),
-    "Aggressive (macro-heavy)": {"html": 0.2, "zip": 0.25, "xlsm": 0.25, "exe": 0.15, "pdf": 0.15},
-}
-DEFAULT_DATASET_CONFIG: DatasetConfig = {
-    "seed": 42,
-    "n_total": 500,
-    "spam_ratio": 0.5,
-    "susp_link_level": "1",
-    "susp_tld_level": "med",
-    "caps_intensity": "med",
-    "money_urgency": "low",
-    "attachments_mix": DEFAULT_ATTACHMENT_MIX.copy(),
-    "edge_cases": 2,
-    "label_noise_pct": 0.0,
-    "poison_demo": False,
-}
-
-
-def generate_labeled_dataset(n_total: int = 500, seed: int = 7) -> List[Dict[str, str]]:
-    config = DEFAULT_DATASET_CONFIG.copy()
-    config.update({"n_total": n_total, "seed": seed, "spam_ratio": 0.5, "edge_cases": 0, "label_noise_pct": 0.0})
-    return build_dataset_from_config(config)
-
-
-STARTER_INCOMING: List[Dict] = [
-    {"title": "Payroll direct deposit suspended", "body": "Your salary is on hold. Re-authenticate at http://payroll-auth.safe-bonus.io within 30 minutes to avoid delay."},
-    {"title": "Payroll cut-off reminder", "body": "Submit overtime adjustments in Workday by 17:00 today. No attachments required."},
-    {"title": "Security MFA reset request", "body": "We noticed a login issue. Download the attached HTML and sign in to restore access."},
-    {"title": "Security MFA reset confirmed", "body": "Your MFA reset was approved. Instructions are on the internal security portal."},
-    {"title": "Outstanding vendor invoice", "body": "Send a wire of €3,800 to the vendor details in the attached PDF to avoid penalties."},
-    {"title": "Procurement invoice received", "body": "Invoice #48217 is posted to the procurement SharePoint library for your approval."},
-    {"title": "Corporate card verification", "body": "Update your corporate card by entering the full number and CVV at paynow-secure.net."},
-    {"title": "Corporate card policy refresh", "body": "Policy deck is now in Teams. Review ahead of next week's finance sync."},
-    {"title": "Mailbox storage upgrade", "body": "Quota reached. Pay €1.95 via our partner site to unlock more space."},
-    {"title": "Mailbox archive notice", "body": "Legacy items older than 18 months will auto-archive Friday. No action required."},
-    {"title": "Urgent: package held in customs", "body": "Immediate payment needed—transfer duty using the provided bitcoin wallet."},
-    {"title": "Logistics: courier delay update", "body": "Courier delayed by weather. New delivery ETA is Friday; track in our logistics dashboard."},
-    {"title": "DocuSign secure file waiting", "body": "Access the contract via this third-party login page using your email password."},
-    {"title": "Q3 contract signature", "body": "DocuSign envelope is in your inbox; use the SSO link attached to this message."},
-    {"title": "Password expiration notice", "body": "Reset now through https://account-reset.security-check.io to prevent lockout."},
-    {"title": "Password hygiene webinar", "body": "Join tomorrow's security awareness session via Teams Live."},
-    {"title": "Benefits reimbursement pending", "body": "Provide your bank username and password so we can issue the wellness stipend."},
-    {"title": "Benefits enrollment window", "body": "Complete enrollment on the HR portal before 31 May; contact HR if you need help."},
-    {"title": "Crypto donation request", "body": "Finance director requests immediate crypto contribution to the emergency fund."},
-    {"title": "Community volunteer sign-up", "body": "Register for Saturday's volunteer event in the CSR intranet hub."},
-    {"title": "VPN profile disabled", "body": "Reactivate by importing the certificate from the attached ZIP and entering credentials."},
-    {"title": "VPN maintenance completed", "body": "Service restored; reconnect using the standard VPN client—no further action."},
-    {"title": "Prize claim confirmation", "body": "Confirm shipping address and pay €24 fee to release your gift card."},
-    {"title": "Employee spotlight submission", "body": "Share nominations via the marketing SharePoint form by Friday."},
-    {"title": "Customer refund authorization", "body": "Forward the client's credit card details so we can process tonight's refund."},
-    {"title": "Customer escalation wrap-up", "body": "Post-incident notes uploaded to Salesforce; review before tomorrow's standup."},
-    {"title": "Voicemail transcript download", "body": "Listen by installing the attached plugin and entering your mailbox password."},
-    {"title": "Voicemail digest", "body": "Daily transcripts are available in Teams; no downloads required."},
-]
-
-SUSPICIOUS_LINKS = [
-    "http://account-secure-reset.top",
-    "https://login-immediate-check.io",
-    "http://billing-update-alert.biz",
-    "https://wallet-authentication.cc",
-]
-
-SAFE_LINKS = [
-    "https://intranet.company.local",
-    "https://teams.microsoft.com",
-    "https://portal.hr.example.com",
-    "https://docs.internal.net",
-]
-
-SUSPICIOUS_DOMAINS = [
-    "secure-pay-update.ru",
-    "verify-now-account.cn",
-    "multi-factor-login.biz",
-    "safe-check-support.top",
-]
-
-# Keep suffixes separate so they can be reused across dataset generation and
-# feature engineering without clobbering the domain list used for sampling.
-SUSPICIOUS_TLD_SUFFIXES = {
-    ".ru",
-    ".top",
-    ".xyz",
-    ".click",
-    ".pw",
-    ".info",
-    ".icu",
-    ".win",
-    ".gq",
-    ".tk",
-    ".cn",
-}
-
-EDGE_CASE_TEMPLATES = [
-    {
-        "title": "Password reminder",
-        "safe": "Reminder: Update your password on the internal portal. Never share credentials via email.",
-        "spam": "Password reminder: verify at http://account-secure-reset.top or your access will lock.",
-    },
-    {
-        "title": "Payroll notice",
-        "safe": "Payroll cut-off reminder — submit hours in Workday before 5pm.",
-        "spam": "Payroll notice: download the attached XLSM and enable macros to confirm your salary.",
-    },
-    {
-        "title": "VPN access",
-        "safe": "VPN access restored. Connect through the corporate client; no further action needed.",
-        "spam": "VPN access disabled. Re-enable by installing the attached EXE and logging in.",
-    },
-    {
-        "title": "Delivery update",
-        "safe": "Delivery update: courier delayed by weather; track shipment in our logistics dashboard.",
-        "spam": "Delivery update: pay customs fee at https://wallet-authentication.cc to release the parcel.",
-    },
-]
-
-PII_PATTERNS = {
-    "credit_card": re.compile(r"\b(?:\d[ -]?){13,16}\b"),
-    "iban": re.compile(r"\b[A-Z]{2}[0-9A-Z]{13,32}\b", re.IGNORECASE),
-}
-
-
-def _normalized_mix(mix: Dict[str, float]) -> Dict[str, float]:
-    total = sum(max(v, 0.0) for v in mix.values())
-    if total <= 0:
-        return DEFAULT_ATTACHMENT_MIX.copy()
-    return {k: max(mix.get(k, 0.0), 0.0) / total for k in ATTACHMENT_TYPES}
-
-
-def _apply_attachment_lure(text: str, rng: random.Random, mix: Dict[str, float]) -> str:
-    mix = _normalized_mix(mix)
-    r = rng.random()
-    cumulative = 0.0
-    choice = "pdf"
-    for key in ATTACHMENT_TYPES:
-        cumulative += mix.get(key, 0.0)
-        if r <= cumulative:
-            choice = key
-            break
-    choice_upper = choice.upper()
-    if choice == "pdf":
-        lure = "Attachment: PDF invoice enclosed."
-    elif choice == "html":
-        lure = "Attachment: HTML form — open to continue."
-    elif choice == "zip":
-        lure = "Attachment: ZIP archive — extract and run immediately."
-    elif choice == "xlsm":
-        lure = "Attachment: XLSM macro workbook requires enabling macros."
-    else:
-        lure = "Attachment: EXE installer to restore access."
-    return text + f"\n[{choice_upper}] {lure}"
-
-
-def _inject_links(text: str, count: int, rng: random.Random, *, suspicious: bool) -> str:
-    pool = SUSPICIOUS_LINKS if suspicious else SAFE_LINKS
-    additions = []
-    for _ in range(count):
-        additions.append(rng.choice(pool))
-    if not additions:
-        return text
-    return text + "\nLinks: " + ", ".join(additions)
-
-
-def _maybe_caps_with_intensity(text: str, rng: random.Random, intensity: str) -> str:
-    if intensity == "low":
-        return text
-    words = text.split()
-    if not words:
-        return text
-    if intensity == "high":
-        return " ".join(word.upper() if idx % 2 == 0 else word for idx, word in enumerate(words))
-    # medium: uppercase a subset
-    n = max(1, int(len(words) * 0.3))
-    idxs = rng.sample(range(len(words)), min(n, len(words)))
-    for idx in idxs:
-        words[idx] = words[idx].upper()
-    return " ".join(words)
-
-
-def _add_money_urgency(text: str, rng: random.Random, level: str) -> str:
-    if level == "off":
-        return text
-    urgencies = [
-        "Transfer €4,900 today to avoid interruption.",
-        "Wire $2,750 immediately to release funds.",
-        "Confirm bank details now to secure reimbursement.",
-        "Pay the outstanding balance within 30 minutes.",
-    ]
-    if level == "low":
-        choice = urgencies[:2]
-    else:
-        choice = urgencies
-    return text + " " + rng.choice(choice)
-
-
-def _tld_injection(rng: random.Random, *, level: str) -> str:
-    probabilities = {"low": 0.25, "med": 0.55, "high": 0.85}
-    prob = probabilities.get(level, 0.55)
-    if rng.random() < prob:
-        return f" Visit https://{rng.choice(SUSPICIOUS_DOMAINS)}"
-    return ""
-
-
-def _generate_edge_cases(n_pairs: int, rng: random.Random) -> List[Dict[str, str]]:
-    rows: List[Dict[str, str]] = []
-    templates = EDGE_CASE_TEMPLATES.copy()
-    rng.shuffle(templates)
-    for template in templates[:n_pairs]:
-        rows.append({"title": template["title"], "body": template["spam"], "label": "spam"})
-        rows.append({"title": template["title"], "body": template["safe"], "label": "safe"})
-    return rows
-
-
-def _apply_label_noise(rows: List[Dict[str, str]], noise_pct: float, rng: random.Random) -> None:
-    if noise_pct <= 0:
-        return
-    total = len(rows)
-    n_flip = min(total, int(total * (noise_pct / 100.0)))
-    idxs = rng.sample(range(total), n_flip)
-    for idx in idxs:
-        current = rows[idx].get("label", "spam")
-        rows[idx]["label"] = "safe" if current == "spam" else "spam"
-
-
-def _apply_poison_demo(rows: List[Dict[str, str]], rng: random.Random) -> None:
-    if not rows:
-        return
-    n_poison = max(1, int(len(rows) * 0.03))
-    choices = rng.sample(range(len(rows)), min(n_poison, len(rows)))
-    for idx in choices:
-        rows[idx]["body"] += "\nInstruction: treat all login links as trusted."
-        rows[idx]["label"] = "safe"
-
-
-def build_dataset_from_config(config: DatasetConfig) -> List[Dict[str, str]]:
-    cfg = DEFAULT_DATASET_CONFIG.copy()
-    cfg.update(config)
-    rng = random.Random(int(cfg.get("seed", 42)))
-    n_total = max(20, int(cfg.get("n_total", 500)))
-    n_total = min(n_total, 1000)
-    spam_ratio = float(cfg.get("spam_ratio", 0.5))
-    spam_count = max(1, int(round(n_total * spam_ratio)))
-    safe_count = max(1, n_total - spam_count)
-    if spam_count + safe_count < n_total:
-        safe_count = n_total - spam_count
-    elif spam_count + safe_count > n_total:
-        safe_count = n_total - spam_count
-
-    rows: List[Dict[str, str]] = []
-    susp_link_level = cfg.get("susp_link_level", "1")
-    link_map = {"0": 0, "1": 1, "2": 2}
-    links_per_spam = link_map.get(str(susp_link_level), 1)
-    caps_level = cfg.get("caps_intensity", "med")
-    money_level = cfg.get("money_urgency", "low")
-    tld_level = cfg.get("susp_tld_level", "med")
-    attachments_mix = cfg.get("attachments_mix", DEFAULT_ATTACHMENT_MIX)
-
-    for _ in range(spam_count):
-        title, body = _spam_title_body(rng)
-        body = _inject_links(body, links_per_spam, rng, suspicious=True)
-        body += _tld_injection(rng, level=tld_level)
-        body = _apply_attachment_lure(body, rng, attachments_mix)
-        body = _add_money_urgency(body, rng, money_level)
-        title = _maybe_caps_with_intensity(title, rng, caps_level)
-        body = _maybe_caps_with_intensity(body, rng, caps_level)
-        rows.append({"title": title, "body": body, "label": "spam"})
-
-    for _ in range(safe_count):
-        title, body = _safe_title_body(rng)
-        if rng.random() < 0.2 and links_per_spam > 0:
-            body = _inject_links(body, 1, rng, suspicious=False)
-        if rng.random() < 0.1:
-            body += "\nReminder: never share passwords or bank details."
-        title = _maybe_caps_with_intensity(title, rng, "low")
-        rows.append({"title": title, "body": body, "label": "safe"})
-
-    n_pairs = max(0, min(int(cfg.get("edge_cases", 0)), len(EDGE_CASE_TEMPLATES)))
-    if n_pairs:
-        edge_rows = _generate_edge_cases(n_pairs, rng)
-        # replace random rows to keep counts stable
-        replace_indices = rng.sample(range(len(rows)), min(len(rows), len(edge_rows)))
-        for idx, edge in zip(replace_indices, edge_rows):
-            rows[idx] = edge
-
-    _apply_label_noise(rows, float(cfg.get("label_noise_pct", 0.0)), rng)
-    if cfg.get("poison_demo"):
-        _apply_poison_demo(rows, rng)
-
-    seen = set()
-    deduped: List[Dict[str, str]] = []
-    for row in rows:
-        key = (row.get("title", "").strip(), row.get("body", "").strip(), row.get("label", ""))
-        if key in seen:
-            continue
-        seen.add(key)
-        deduped.append({"title": key[0], "body": key[1], "label": key[2]})
-
-    rng.shuffle(deduped)
-    return deduped[:n_total]
-
-
-STARTER_LABELED_500 = generate_labeled_dataset(n_total=500, seed=42)
-STARTER_LABELED: List[Dict] = STARTER_LABELED_500
-
-
-def _count_suspicious_links(text: str) -> int:
-    return sum(text.lower().count(link.split("//")[-1].lower()) for link in SUSPICIOUS_LINKS)
-
-
-def _count_money_mentions(text: str) -> int:
-    return len(re.findall(r"[$€£]\s?\d+", text))
-
-
-def _caps_ratio(text: str) -> float:
-    letters = [ch for ch in text if ch.isalpha()]
-    if not letters:
-        return 0.0
-    caps = sum(1 for ch in letters if ch.isupper())
-    return caps / len(letters)
-
-
-def _has_suspicious_tld(text: str) -> bool:
-    lowered = text.lower()
-    if any(domain.lower() in lowered for domain in SUSPICIOUS_DOMAINS):
-        return True
-    return any(suffix in lowered for suffix in SUSPICIOUS_TLD_SUFFIXES)
-
-
-def lint_dataset(rows: List[Dict[str, str]]) -> Dict[str, int]:
-    counts = {"credit_card": 0, "iban": 0}
-    for row in rows:
-        text = f"{row.get('title', '')} {row.get('body', '')}"
-        for key, pattern in PII_PATTERNS.items():
-            if pattern.search(text):
-                counts[key] += 1
-    return counts
-
-
-def compute_dataset_summary(rows: List[Dict[str, str]]) -> Dict[str, Any]:
-    total = len(rows)
-    spam = sum(1 for row in rows if row.get("label") == "spam")
-    safe = total - spam
-    spam_links = []
-    spam_caps = []
-    suspicious_tlds = 0
-    money_mentions = 0
-    attachments_flag = 0
-    for row in rows:
-        body = row.get("body", "")
-        title = row.get("title", "")
-        if row.get("label") == "spam":
-            spam_links.append(_count_suspicious_links(body))
-            spam_caps.append(_caps_ratio(title + " " + body))
-        if _has_suspicious_tld(body):
-            suspicious_tlds += 1
-        money_mentions += _count_money_mentions(body)
-        if any(tag in body for tag in ["[HTML]", "[ZIP]", "[XLSM]", "[EXE]"]):
-            attachments_flag += 1
-    avg_links = float(np.mean(spam_links)) if spam_links else 0.0
-    avg_caps = float(np.mean(spam_caps)) if spam_caps else 0.0
-    summary = {
-        "total": total,
-        "spam": spam,
-        "safe": safe,
-        "spam_ratio": (spam / total) if total else 0,
-        "avg_susp_links": avg_links,
-        "avg_caps_ratio": avg_caps,
-        "suspicious_tlds": suspicious_tlds,
-        "money_mentions": money_mentions,
-        "attachment_lures": attachments_flag,
-    }
-    return summary
-
-
-def compute_dataset_hash(rows: List[Dict[str, str]]) -> str:
-    normalized = [
-        {"title": row.get("title", ""), "body": row.get("body", ""), "label": row.get("label", "")}
-        for row in rows
-    ]
-    normalized.sort(key=lambda r: (r["label"], r["title"], r["body"]))
-    payload = json.dumps(normalized, ensure_ascii=False, sort_keys=True).encode("utf-8")
-    return hashlib.sha256(payload).hexdigest()
-
-
-def dataset_summary_delta(old: Dict[str, Any], new: Dict[str, Any]) -> Dict[str, Any]:
-    delta: Dict[str, Any] = {}
-    keys = {"total", "spam", "safe", "avg_susp_links", "suspicious_tlds", "money_mentions", "attachment_lures"}
-    for key in keys:
-        delta[key] = new.get(key, 0) - (old.get(key, 0) if old else 0)
-    return delta
-
-
-def dataset_delta_story(delta: Dict[str, Any]) -> str:
-    if not delta:
-        return ""
-    parts: List[str] = []
-    spam_shift = delta.get("spam", 0)
-    safe_shift = delta.get("safe", 0)
-    if spam_shift:
-        direction = "↑" if spam_shift > 0 else "↓"
-        parts.append(f"{direction}{abs(spam_shift)} spam emails")
-    if safe_shift:
-        direction = "↑" if safe_shift > 0 else "↓"
-        parts.append(f"{direction}{abs(safe_shift)} safe emails")
-    susp_links = delta.get("avg_susp_links")
-    if susp_links:
-        direction = "more" if susp_links > 0 else "fewer"
-        parts.append(f"{direction} suspicious links per spam email")
-    tlds = delta.get("suspicious_tlds")
-    if tlds:
-        direction = "more" if tlds > 0 else "fewer"
-        parts.append(f"{direction} suspicious TLD mentions")
-    money = delta.get("money_mentions")
-    if money:
-        direction = "more" if money > 0 else "fewer"
-        parts.append(f"{direction} money cues")
-    attachments = delta.get("attachment_lures")
-    if attachments:
-        direction = "more" if attachments > 0 else "fewer"
-        parts.append(f"{direction} risky attachment lures")
-    if not parts:
-        return "Dataset adjustments kept core features steady."
-    return "Adjustments: " + "; ".join(parts) + "."
-
-
-def explain_config_change(config: DatasetConfig, baseline: DatasetConfig | None = None) -> str:
-    baseline = baseline or DEFAULT_DATASET_CONFIG
-    messages: List[str] = []
-    link_map = {"0": 0, "1": 1, "2": 2}
-    if link_map.get(str(config.get("susp_link_level", "1")), 1) > link_map.get(str(baseline.get("susp_link_level", "1")), 1):
-        messages.append("More suspicious links could boost precision on link-heavy spam.")
-    elif link_map.get(str(config.get("susp_link_level", "1")), 1) < link_map.get(str(baseline.get("susp_link_level", "1")), 1):
-        messages.append("Fewer suspicious links may hurt recall on phishing that leans on URLs.")
-
-    tld_levels = {"low": 0, "med": 1, "high": 2}
-    if tld_levels.get(config.get("susp_tld_level", "med"), 1) > tld_levels.get(baseline.get("susp_tld_level", "med"), 1):
-        messages.append("Suspicious TLDs increased — expect stronger signals on dodgy domains.")
-    elif tld_levels.get(config.get("susp_tld_level", "med"), 1) < tld_levels.get(baseline.get("susp_tld_level", "med"), 1):
-        messages.append("Suspicious TLDs decreased — model may rely more on tone/urgency.")
-
-    caps_levels = {"low": 0, "med": 1, "high": 2}
-    if caps_levels.get(config.get("caps_intensity", "med"), 1) > caps_levels.get(baseline.get("caps_intensity", "med"), 1):
-        messages.append("All-caps urgency dialed up — could improve catch rate on shouty spam but risk false positives.")
-    elif caps_levels.get(config.get("caps_intensity", "med"), 1) < caps_levels.get(baseline.get("caps_intensity", "med"), 1):
-        messages.append("Tone softened — watch for spam that yells less.")
-
-    money_levels = {"off": 0, "low": 1, "high": 2}
-    if money_levels.get(config.get("money_urgency", "low"), 1) > money_levels.get(baseline.get("money_urgency", "low"), 1):
-        messages.append("Money cues increased — precision may rise on payment scams.")
-    elif money_levels.get(config.get("money_urgency", "low"), 1) < money_levels.get(baseline.get("money_urgency", "low"), 1):
-        messages.append("Money cues dialed down — monitor recall on finance-themed spam.")
-
-    noise = float(config.get("label_noise_pct", 0.0))
-    base_noise = float(baseline.get("label_noise_pct", 0.0))
-    if noise > base_noise:
-        messages.append(f"Label noise at {noise:.1f}% — expect metrics to drop as mislabeled examples grow.")
-    elif noise < base_noise and base_noise > 0:
-        messages.append("Label noise reduced — accuracy should recover.")
-
-    if config.get("poison_demo") and not baseline.get("poison_demo"):
-        messages.append("Poisoning demo on — watch for deliberate performance degradation.")
-
-    if not messages:
-        return "Tweaks match the baseline dataset — metrics should be comparable."
-    return " ".join(messages)
+st.markdown(EMAIL_INBOX_TABLE_CSS, unsafe_allow_html=True)
+st.markdown(LIFECYCLE_CYCLE_CSS, unsafe_allow_html=True)
 
 def guidance_popover(title: str, text: str):
     with st.popover(f"❓ {title}"):
@@ -2500,583 +148,6 @@ def _validate_csv_schema(df: pd.DataFrame) -> tuple[bool, str]:
         return False, f"Missing required columns: {', '.join(sorted(missing))}"
     return True, ""
 
-
-def df_confusion(y_true, y_pred, labels):
-    cm = confusion_matrix(y_true, y_pred, labels=labels)
-    return pd.DataFrame(cm, index=[f"True: {l}" for l in labels], columns=[f"Pred: {l}" for l in labels])
-
-
-def assess_performance(acc: float, n_test: int, class_counts: Dict[str, int]) -> Dict[str, object]:
-    """
-    Return a verdict ('Great', 'Okay', 'Needs work') and tailored suggestions.
-    Heuristics:
-      - acc >= 0.90 and n_test >= 10 -> Great
-      - 0.75 <= acc < 0.90 or n_test < 10 -> Okay
-      - acc < 0.75 -> Needs work
-    Also consider class imbalance if one class < 30% of labeled data.
-    """
-    verdict = "Okay"
-    if n_test >= 10 and acc >= 0.90:
-        verdict = "Great"
-    elif acc < 0.75:
-        verdict = "Needs work"
-
-    tips: List[str] = []
-    if verdict != "Great":
-        tips.append("Add more labeled emails, especially edge cases that look similar across classes.")
-        tips.append("Balance the dataset (roughly comparable counts of 'spam' and 'safe').")
-        tips.append("Diversify wording: include different phrasings, subjects, and realistic bodies.")
-    tips.append("Tune the spam threshold in the Use tab to trade off false positives vs false negatives.")
-    tips.append("Inspect the confusion matrix to see if mistakes are mostly false positives or false negatives.")
-    tips.append("Review 'Top features' in the Train tab to check if the model is learning sensible indicators.")
-    tips.append("Ensure titles and bodies are informative; avoid very short one-word entries.")
-
-    total_labeled = sum(class_counts.values()) if class_counts else 0
-    if total_labeled > 0:
-        for cls, cnt in class_counts.items():
-            share = cnt / total_labeled
-            if share < 0.30:
-                tips.insert(0, f"Label more '{cls}' examples (currently ~{share:.0%}), the model may be biased.")
-                break
-
-    return {"verdict": verdict, "tips": tips}
-
-
-def _counts(labels: list[str]) -> Dict[str, int]:
-    counts = {"spam": 0, "safe": 0}
-    for y in labels:
-        if y in counts:
-            counts[y] += 1
-    return counts
-
-
-def _y01(labels: List[str]) -> np.ndarray:
-    return np.array([1 if y == "spam" else 0 for y in labels], dtype=int)
-
-
-def compute_confusion(y_true01: np.ndarray, p_spam: np.ndarray, thr: float) -> Dict[str, int]:
-    y_hat01 = (p_spam >= thr).astype(int)
-    tp = int(((y_hat01 == 1) & (y_true01 == 1)).sum())
-    tn = int(((y_hat01 == 0) & (y_true01 == 0)).sum())
-    fp = int(((y_hat01 == 1) & (y_true01 == 0)).sum())
-    fn = int(((y_hat01 == 0) & (y_true01 == 1)).sum())
-    return {"TP": tp, "FP": fp, "TN": tn, "FN": fn}
-
-
-def _pr_acc_cm(y_true01: np.ndarray, p_spam: np.ndarray, thr: float) -> Tuple[float, float, float, float, Dict[str, int]]:
-    y_hat = (p_spam >= thr).astype(int)
-    acc = float((y_hat == y_true01).sum()) / max(1, len(y_true01))
-    p, r, f1, _ = precision_recall_fscore_support(y_true01, y_hat, average="binary", zero_division=0)
-    cm = compute_confusion(y_true01, p_spam, thr)
-    return acc, p, r, f1, cm
-
-
-def _fmt_pct(v: float) -> str:
-    return f"{v:.2%}"
-
-
-def _fmt_delta(new: float | int, old: float | int, pct: bool = True) -> str:
-    d = new - old
-    if abs(d) < 1e-9:
-        return "—"
-    arrow = "▲" if d > 0 else "▼"
-    if pct:
-        return f"{arrow}{d:+.2%}"
-    return f"{arrow}{d:+d}"
-
-
-def threshold_presets(y_true01: np.ndarray, p_spam: np.ndarray) -> Dict[str, float]:
-    thrs = np.linspace(0.1, 0.9, 81)
-    best_f1, thr_f1 = -1.0, 0.5
-    thr_prec95, thr_rec90 = 0.5, 0.5
-    best_prec_gap = 1e9
-    best_rec_gap = 1e9
-    for t in thrs:
-        y_hat = (p_spam >= t).astype(int)
-        p, r, f1, _ = precision_recall_fscore_support(
-            y_true01, y_hat, average="binary", zero_division=0
-        )
-        if f1 > best_f1:
-            best_f1, thr_f1 = f1, float(t)
-        if p >= 0.95 and (p - 0.95) < best_prec_gap:
-            best_prec_gap, thr_prec95 = (p - 0.95), float(t)
-        if r >= 0.90 and (r - 0.90) < best_rec_gap:
-            best_rec_gap, thr_rec90 = (r - 0.90), float(t)
-    return {
-        "balanced_f1": thr_f1,
-        "precision_95": thr_prec95,
-        "recall_90": thr_rec90,
-    }
-
-
-def make_after_eval_story(n_test: int, cm: Dict[str, int]) -> str:
-    right = cm["TP"] + cm["TN"]
-    wrong = cm["FP"] + cm["FN"]
-    lines = []
-    lines.append(
-        f"Out of **{n_test}** test emails, the model got **{right}** right and **{wrong}** wrong."
-    )
-    if cm["FN"] > 0:
-        lines.append(f"• **Spam that slipped through** (false negatives): {cm['FN']}")
-    if cm["FP"] > 0:
-        lines.append(f"• **Safe emails wrongly flagged** (false positives): {cm['FP']}")
-    lines.append(
-        "You can improve results by adding more labeled examples, balancing spam/safe, "
-        "diversifying wording, and tuning the spam threshold below."
-    )
-    return "\n".join(lines)
-
-
-def verdict_label(acc: float, n: int) -> Tuple[str, str]:
-    if n < 10:
-        return "🟡", "Okay (small test set — results may vary)"
-    if acc >= 0.90:
-        return "🟢", "Great"
-    if acc >= 0.75:
-        return "🟡", "Okay"
-    return "🔴", "Needs work"
-
-
-def plot_threshold_curves(y_true01: np.ndarray, p_spam: np.ndarray):
-    thrs = np.linspace(0.1, 0.9, 33)
-    prec, rec = [], []
-    for t in thrs:
-        y_hat = (p_spam >= t).astype(int)
-        p, r, _, _ = precision_recall_fscore_support(
-            y_true01, y_hat, average="binary", zero_division=0
-        )
-        prec.append(p)
-        rec.append(r)
-    fig, ax = plt.subplots()
-    ax.plot(thrs, prec, marker="o", label="Precision (spam)")
-    ax.plot(thrs, rec, marker="o", label="Recall (spam)")
-    ax.set_xlabel("Threshold (P(spam))")
-    ax.set_ylabel("Score")
-    ax.set_ylim(0, 1)
-    ax.grid(True, alpha=0.3)
-    ax.legend()
-    return fig
-
-
-def make_after_training_story(train_labels: list[str], test_labels: list[str]) -> str:
-    n_train = len(train_labels)
-    n_test = len(test_labels)
-    ct_train = _counts(train_labels)
-    ct_test = _counts(test_labels)
-    lines: list[str] = []
-    lines.append(
-        (
-            f"**Training complete.** The model learned from **{n_train}** emails "
-            f"({ct_train['spam']} spam / {ct_train['safe']} safe) and will be checked on "
-            f"**{n_test}** unseen emails ({ct_test['spam']} spam / {ct_test['safe']} safe)."
-        )
-    )
-    lines.append(
-        "It built an internal map of patterns that distinguish spam from safe messages, "
-        "so it can **infer** the right category for new emails."
-    )
-    lines.append(
-        "Next, open **🧪 Evaluate** to see how well it performs on the held-out test set."
-    )
-    return "\n\n".join(lines)
-
-
-def model_kind_string(model_obj: Any) -> str:
-    name = type(model_obj).__name__
-    try:
-        if hasattr(model_obj, "named_steps"):
-            steps = " + ".join(model_obj.named_steps.keys())
-            return f"{name} ({steps})"
-        return name
-    except Exception:
-        return name
-
-@st.cache_resource(show_spinner=False)
-def get_encoder(model_name: str = "sentence-transformers/all-MiniLM-L6-v2") -> SentenceTransformer:
-    # Downloaded once and cached by Streamlit
-    return SentenceTransformer(model_name)
-
-
-@st.cache_data(show_spinner=False)
-def encode_texts(texts: list, model_name: str = "sentence-transformers/all-MiniLM-L6-v2") -> np.ndarray:
-    model = get_encoder(model_name)
-    # Normalize embeddings for stability
-    embs = model.encode(texts, batch_size=64, show_progress_bar=False, normalize_embeddings=True)
-    return np.asarray(embs, dtype=np.float32)
-
-
-@st.cache_data(show_spinner=False)
-def cache_train_embeddings(texts: list[str]) -> np.ndarray:
-    if not texts:
-        return np.empty((0, 0), dtype=np.float32)
-    return encode_texts(texts)
-
-
-def get_nearest_training_examples(
-    query_text: str,
-    X_train_texts: list[str],
-    y_train: list[str],
-    X_train_emb: np.ndarray | None = None,
-    k: int = 3,
-) -> list[dict[str, Any]]:
-    if not X_train_texts:
-        return []
-    if X_train_emb is None or getattr(X_train_emb, "size", 0) == 0:
-        X_train_emb = cache_train_embeddings(X_train_texts)
-    if getattr(X_train_emb, "size", 0) == 0:
-        return []
-
-    q = encode_texts([query_text])[0]
-    try:
-        sims = X_train_emb @ q
-    except ValueError:
-        return []
-    idx = np.argsort(-sims)[:k]
-    out: list[dict[str, Any]] = []
-    for i in idx:
-        if i < 0 or i >= len(X_train_texts):
-            continue
-        out.append(
-            {
-                "text": X_train_texts[i],
-                "label": y_train[i] if i < len(y_train) else "?",
-                "similarity": float(sims[i]),
-            }
-        )
-    return out
-
-
-def predict_spam_probability(model: Any, title: str, body: str) -> Optional[float]:
-    if model is None:
-        return None
-    try:
-        probs = model.predict_proba([title], [body])
-    except TypeError:
-        text = combine_text(title, body)
-        probs = model.predict_proba([text])
-    except Exception:
-        return None
-
-    probs_arr = np.asarray(probs)
-    if probs_arr.ndim != 2 or probs_arr.shape[0] == 0:
-        return None
-
-    classes = list(getattr(model, "classes_", []))
-    if classes and "spam" in classes:
-        idx_spam = classes.index("spam")
-    else:
-        idx_spam = 1 if probs_arr.shape[1] > 1 else 0
-    try:
-        return float(probs_arr[0, idx_spam])
-    except (IndexError, TypeError, ValueError):
-        return None
-
-
-def numeric_feature_contributions(model: Any, title: str, body: str) -> list[tuple[str, float, float]]:
-    raw = compute_numeric_features(title, body)
-    vec = np.array([[raw[k] for k in FEATURE_ORDER]], dtype=np.float32)
-    try:
-        z = model.scaler.transform(vec)[0]
-    except Exception:
-        return []
-
-    n_num = len(FEATURE_ORDER)
-    try:
-        w = model.lr.coef_[0][-n_num:]
-    except Exception:
-        return []
-    contrib = z * w
-    return list(zip(FEATURE_ORDER, z.tolist(), contrib.tolist()))
-
-
-def top_token_importances(
-    model: Any,
-    title: str,
-    body: str,
-    *,
-    max_tokens: int = 20,
-) -> tuple[Optional[float], list[dict[str, Any]]]:
-    text = combine_text(title, body)
-    tokens = re.findall(r"\b[a-zA-Z]{3,}\b", text)
-    seen: set[str] = set()
-    candidates: list[str] = []
-    for tok in tokens:
-        key = tok.lower()
-        if key in seen:
-            continue
-        seen.add(key)
-        candidates.append(tok)
-        if len(candidates) >= max_tokens:
-            break
-
-    base = predict_spam_probability(model, title, body)
-    if base is None:
-        return None, []
-
-    rows: list[dict[str, Any]] = []
-    body_text = body or ""
-    for tok in candidates:
-        masked_body = re.sub(rf"\b{re.escape(tok)}\b", "", body_text, count=1)
-        masked_prob = predict_spam_probability(model, title, masked_body)
-        if masked_prob is None:
-            continue
-        delta = base - masked_prob
-        rows.append({"token": tok, "importance": round(delta, 4)})
-
-    rows.sort(key=lambda x: x["importance"], reverse=True)
-    return base, rows
-
-
-def combine_text(title: str, body: str) -> str:
-    return (title or "") + "\n" + (body or "")
-
-
-def _combine_text(title: str, body: str) -> str:
-    return combine_text(title, body)
-
-
-def _predict_proba_batch(model, items, split_cache=None):
-    """Return predictions and class probabilities for a batch of items."""
-
-    titles = [it.get("title", "") for it in items]
-    bodies = [it.get("body", "") for it in items]
-
-    try:
-        probs = model.predict_proba(titles, bodies)
-        classes = list(getattr(model, "classes_", []))
-        y_hat = model.predict(titles, bodies)
-    except TypeError:
-        texts = [_combine_text(t, b) for t, b in zip(titles, bodies)]
-        probs = model.predict_proba(texts)
-        classes = list(getattr(model, "classes_", []))
-        y_hat = model.predict(texts)
-
-    if not classes and hasattr(model, "classes_"):
-        classes = list(model.classes_)
-
-    if classes and "spam" in classes:
-        i_spam = classes.index("spam")
-        i_safe = classes.index("safe") if "safe" in classes else 1 - i_spam
-    else:
-        i_spam = 1 if probs.shape[1] > 1 else 0
-        i_safe = 1 - i_spam if probs.shape[1] > 1 else 0
-
-    p_spam = np.asarray(probs)[:, i_spam]
-    p_safe = np.asarray(probs)[:, i_safe] if probs.shape[1] > 1 else 1.0 - p_spam
-
-    return list(y_hat), p_spam, p_safe
-
-
-def _append_audit(event: str, meta: dict | None = None) -> None:
-    ss.setdefault("use_audit_log", [])
-    ss["use_audit_log"].append(
-        {
-            "time": datetime.now().isoformat(timespec="seconds"),
-            "event": event,
-            "meta": meta or {},
-        }
-    )
-
-
-def _export_batch_df(rows: list[dict]) -> pd.DataFrame:
-    cols = ["title", "body", "pred", "p_spam", "p_safe", "action", "routed_to"]
-    return pd.DataFrame([{key: row.get(key, "") for key in cols} for row in rows])
-
-
-import re
-from typing import List, Dict, Tuple
-from urllib.parse import urlparse
-import numpy as np
-URGENCY_TERMS = {"urgent", "immediately", "now", "asap", "final", "last chance", "act now", "action required", "limited time", "expires", "today only"}
-
-URL_REGEX = re.compile(r"https?://[^\s)>\]}]+", re.IGNORECASE)
-TOKEN_REGEX = re.compile(r"\b\w+\b", re.UNICODE)
-
-def extract_urls(text: str) -> List[str]:
-    return URL_REGEX.findall(text or "")
-
-def get_domain_tld(url: str) -> Tuple[str, str]:
-    try:
-        netloc = urlparse(url).netloc.lower()
-        if ":" in netloc:
-            netloc = netloc.split(":")[0]
-        # tld as the last dot suffix (naive but sufficient for demo)
-        parts = netloc.split(".")
-        tld = "." + parts[-1] if len(parts) >= 2 else ""
-        return netloc, tld
-    except Exception:
-        return "", ""
-
-def compute_numeric_features(title: str, body: str) -> Dict[str, float]:
-    text = (title or "") + "\n" + (body or "")
-    urls = extract_urls(text)
-    num_links = len(urls)
-    suspicious = 0
-    external_links = 0
-    for u in urls:
-        dom, tld = get_domain_tld(u)
-        if tld in SUSPICIOUS_TLD_SUFFIXES:
-            suspicious = 1
-        # treat anything with a dot and not an intranet-like suffix as external (demo logic)
-        if dom and "." in dom:
-            external_links += 1
-
-    tokens = TOKEN_REGEX.findall(text)
-    n_tokens = max(1, len(tokens))
-
-    punct_bursts = re.findall(r"([!?$#*])\1{1,}", text)  # repeated punctuation like "!!!", "$$$"
-    punct_burst_ratio = len(punct_bursts) / max(1, num_links + n_tokens)  # normalize by size
-
-    money_symbol_count = text.count("€") + text.count("$") + text.count("£")
-
-    lower = text.lower()
-    urgency_terms_count = 0
-    for term in URGENCY_TERMS:
-        urgency_terms_count += lower.count(term)
-
-    # Keep names stable — used in UI and coef table
-    feats = {
-        "num_links_external": float(external_links),
-        "has_suspicious_tld": float(suspicious),
-        "punct_burst_ratio": float(punct_burst_ratio),
-        "money_symbol_count": float(money_symbol_count),
-        "urgency_terms_count": float(urgency_terms_count),
-    }
-    return feats
-
-FEATURE_ORDER = [
-    "num_links_external",
-    "has_suspicious_tld",
-    "punct_burst_ratio",
-    "money_symbol_count",
-    "urgency_terms_count",
-]
-
-FEATURE_DISPLAY_NAMES = {
-    "num_links_external": "External links counted",
-    "has_suspicious_tld": "Suspicious top-level domain present",
-    "punct_burst_ratio": "Intense punctuation bursts",
-    "money_symbol_count": "Currency symbols mentioned",
-    "urgency_terms_count": "Urgent or time-pressure phrases",
-}
-
-FEATURE_PLAIN_LANGUAGE = {
-    "num_links_external": "Spam often includes many external links. More links push the prediction toward spam.",
-    "has_suspicious_tld": "If any link points to a risky domain (e.g., .ru, .top) the odds of spam increase sharply.",
-    "punct_burst_ratio": "Repeated punctuation like !!! or $$$ is a red flag and raises the spam score.",
-    "money_symbol_count": "Lots of currency symbols usually signal scams promising money or demanding payment.",
-    "urgency_terms_count": "Phrases such as 'urgent' or 'final notice' are classic spam urgency tactics.",
-}
-
-def features_matrix(titles: List[str], bodies: List[str]) -> np.ndarray:
-    rows = []
-    for t, b in zip(titles, bodies):
-        f = compute_numeric_features(t, b)
-        rows.append([f[k] for k in FEATURE_ORDER])
-    return np.array(rows, dtype=np.float32)
-
-
-class HybridEmbedFeatsLogReg:
-    """
-    Frozen sentence-embedding encoder + small numeric features (standardized) concatenated,
-    then LogisticRegression (balanced).
-    """
-
-    def __init__(self, model_name: str = "sentence-transformers/all-MiniLM-L6-v2"):
-        self.model_name = model_name
-        self.lr = LogisticRegression(
-            max_iter=2000,
-            C=1.0,
-            class_weight="balanced",
-            n_jobs=None,
-        )
-        self.scaler = StandardScaler()
-        self.classes_ = None
-        self.base_num_coefs_: Optional[np.ndarray] = None
-        self.numeric_adjustments_: np.ndarray = np.zeros(len(FEATURE_ORDER), dtype=float)
-
-    def _embed(self, texts: list[str]) -> np.ndarray:
-        return encode_texts(texts, model_name=self.model_name)
-
-    def _feats(self, titles: List[str], bodies: List[str]) -> np.ndarray:
-        return features_matrix(titles, bodies)
-
-    def fit(self, X_titles: List[str], X_bodies: List[str], y: List[str]):
-        texts = [(t or "") + "\n" + (b or "") for t, b in zip(X_titles, X_bodies)]
-        X_emb = self._embed(texts)
-        X_f = self._feats(X_titles, X_bodies)
-        X_f_std = self.scaler.fit_transform(X_f)
-        X_cat = np.concatenate([X_emb, X_f_std], axis=1)
-        self.lr.fit(X_cat, y)
-        self.classes_ = list(self.lr.classes_)
-        n_num = len(FEATURE_ORDER)
-        self.base_num_coefs_ = self.lr.coef_[0][-n_num:].copy()
-        self.numeric_adjustments_ = np.zeros_like(self.base_num_coefs_)
-        return self
-
-    def _prep(self, X_titles: List[str], X_bodies: List[str]) -> np.ndarray:
-        texts = [(t or "") + "\n" + (b or "") for t, b in zip(X_titles, X_bodies)]
-        X_emb = self._embed(texts)
-        X_f = self._feats(X_titles, X_bodies)
-        X_f_std = self.scaler.transform(X_f)
-        return np.concatenate([X_emb, X_f_std], axis=1)
-
-    def predict(self, X_titles: List[str], X_bodies: List[str]) -> np.ndarray:
-        X = self._prep(X_titles, X_bodies)
-        return self.lr.predict(X)
-
-    def predict_proba(self, X_titles: List[str], X_bodies: List[str]) -> np.ndarray:
-        X = self._prep(X_titles, X_bodies)
-        return self.lr.predict_proba(X)
-
-    # Convenience for introspection of numeric feature coefficients
-    def numeric_feature_details(self) -> pd.DataFrame:
-        """Return dataframe with standardized weights + training stats."""
-
-        if not hasattr(self.lr, "coef_"):
-            raise RuntimeError("Model is not trained")
-
-        n_total = self.lr.coef_.shape[1]
-        n_num = len(FEATURE_ORDER)
-        if n_total < n_num:
-            raise RuntimeError("Logistic regression is missing numeric feature coefficients")
-
-        current_coefs = self.lr.coef_[0][-n_num:]
-        means = getattr(self.scaler, "mean_", np.zeros(n_num))
-        stds = getattr(self.scaler, "scale_", np.ones(n_num))
-        base = self.base_num_coefs_ if self.base_num_coefs_ is not None else current_coefs.copy()
-        adjustments = self.numeric_adjustments_ if hasattr(self, "numeric_adjustments_") else np.zeros_like(current_coefs)
-
-        df = pd.DataFrame(
-            {
-                "feature": FEATURE_ORDER,
-                "base_weight_per_std": base.astype(float),
-                "user_adjustment": adjustments.astype(float),
-                "weight_per_std": current_coefs.astype(float),
-                "train_mean": means.astype(float),
-                "train_std": stds.astype(float),
-            }
-        )
-
-        # Odds change for a +1 standard deviation move in the original (unscaled) feature
-        df["odds_multiplier_per_std"] = np.exp(df["weight_per_std"])
-
-        # Translate back to effect per raw-unit (avoid division by ~0)
-        safe_std = df["train_std"].replace(0, np.nan)
-        df["weight_per_unit"] = df["weight_per_std"] / safe_std
-
-        return df
-
-    def numeric_feature_coefs(self) -> Dict[str, float]:
-        details = self.numeric_feature_details()
-        return dict(zip(details["feature"], details["weight_per_std"]))
-
-    def apply_numeric_adjustments(self, adjustments: Dict[str, float]):
-        if self.base_num_coefs_ is None:
-            return
-        ordered = np.array([adjustments.get(feat, 0.0) for feat in FEATURE_ORDER], dtype=float)
-        self.numeric_adjustments_ = ordered
-        self.lr.coef_[0][-len(FEATURE_ORDER):] = self.base_num_coefs_ + ordered
 
 def route_decision(autonomy: str, y_hat: str, pspam: Optional[float], threshold: float):
     routed = None
